@@ -1,4 +1,4 @@
-name: qonvay_2
+name: qonvay
 description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,45 +27,66 @@ dependencies:
     sdk: flutter
   auto_size_text: 3.0.0
   built_collection: 5.1.1
-  built_value: 8.1.3
-  cached_network_image: 3.1.0+1
-  cloud_firestore: 2.5.4
-  firebase_auth: 3.2.0
-  firebase_core: 1.10.0
-  firebase_storage: 10.1.0
+  built_value: 8.4.0
+  cached_network_image: 3.2.1
+  cloud_firestore: 3.2.1
+  cloud_firestore_platform_interface: 5.5.10
+  cloud_firestore_web: 2.6.19
+  collection: 1.16.0
+  easy_debounce: 2.0.1
+  file_picker: 4.5.1
+  firebase_analytics: 9.1.12
+  firebase_analytics_platform_interface: 3.1.10
+  firebase_analytics_web: 0.4.0+17
+  firebase_auth: 3.4.1
+  firebase_auth_platform_interface: 6.3.1
+  firebase_auth_web: 3.3.19
+  firebase_core: 1.19.1
+  firebase_core_platform_interface: 4.4.3
+  firebase_core_web: 1.6.6
+  firebase_storage: 10.3.1
+  firebase_storage_platform_interface: 4.1.10
+  firebase_storage_web: 3.2.19
   flutter_credit_card: 2.0.0
-  flutter_spinkit: 4.1.2+1
+  flutter_google_places: 0.3.0
+  flutter_spinkit: 5.1.0
   flutter_svg: 0.22.0
-  font_awesome_flutter: 8.11.0
-  google_fonts: 2.1.0
-  google_maps_flutter: 2.1.1
-  google_maps_flutter_web: 0.3.2
-  google_sign_in: 5.2.1
-  image_picker: 0.8.4+4
+  font_awesome_flutter: 10.1.0
+  google_fonts: 3.0.1
+  google_sign_in: 5.4.0
+  google_sign_in_android: 6.0.1
+  google_sign_in_ios: 5.4.0
+  google_sign_in_platform_interface: 2.2.0
+  google_sign_in_web: 0.10.2
+  image_picker: 0.8.5+3
   intl: 0.17.0
-  json_path: 0.2.1
-  json_serializable: 6.1.3
+  json_path: 0.4.1
+  json_serializable: 6.3.1
+  material: ^1.0.0+2
   mime_type: 1.0.0
-  page_transition: 2.0.4
-  rxdart: 0.26.0
-  sign_in_with_apple:
-    git:
-      url: https://github.com/asmengistu/dart_packages
-      ref: ab18d53a80e0a4bb90f1ad84f37f1b408930f28d
-      path: packages/sign_in_with_apple/sign_in_with_apple/
+  page_transition: 2.0.9
+  plugin_platform_interface: 2.1.2
+  rxdart: 0.27.4
+  shared_preferences: 2.0.15
+  sign_in_with_apple: 4.1.0
+  sign_in_with_apple_platform_interface: 1.0.0
+  sign_in_with_apple_web: 1.0.1
   smooth_page_indicator: 1.0.0+2
-  timeago: 3.1.0
-  url_launcher: 6.0.15
+  timeago: 3.2.2
+  url_launcher: 6.1.5
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
 
+dependency_overrides:
+  rxdart: 0.27.4
+
 dev_dependencies:
-  build_runner: 2.1.7
-  built_value_generator: 8.1.3
-  flutter_launcher_icons: 0.9.2
+  build_runner: 2.2.0
+  built_value_generator: 8.4.0
+  flutter_launcher_icons: 0.9.3
 
   flutter_test:
     sdk: flutter
@@ -97,6 +118,8 @@ flutter:
     - assets/audios/
     - assets/lottie_animations/
     - assets/rive_animations/
+    - assets/pdfs/
+
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
