@@ -1,6 +1,7 @@
 import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import '../flutter_flow/flutter_flow_util.dart';
 
 import '../backend/backend.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
@@ -11,19 +12,22 @@ export 'anonymous_auth.dart';
 export 'apple_auth.dart';
 export 'email_auth.dart';
 export 'google_auth.dart';
+export 'jwt_token_auth.dart';
 
 /// Tries to sign in or create an account using Firebase Auth.
 /// Returns the User object if sign in was successful.
-Future<User> signInOrCreateAccount(
-    BuildContext context, Future<UserCredential> Function() signInFunc) async {
+Future<User?> signInOrCreateAccount(
+    BuildContext context, Future<UserCredential?> Function() signInFunc) async {
   try {
     final userCredential = await signInFunc();
-    await maybeCreateUser(userCredential.user);
-    return userCredential.user;
+    if (userCredential?.user != null) {
+      await maybeCreateUser(userCredential!.user!);
+    }
+    return userCredential?.user;
   } on FirebaseAuthException catch (e) {
     ScaffoldMessenger.of(context).hideCurrentSnackBar();
     ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(content: Text('Error: ${e.message}')),
+      SnackBar(content: Text('Error: ${e.message!}')),
     );
     return null;
   }
@@ -31,34 +35,53 @@ Future<User> signInOrCreateAccount(
 
 Future signOut() {
   _currentJwtToken = '';
-  FirebaseAuth.instance.signOut();
+  return FirebaseAuth.instance.signOut();
 }
 
-Future resetPassword({String email, BuildContext context}) async {
+Future deleteUser(BuildContext context) async {
+  try {
+    if (currentUser?.user == null) {
+      print('Error: delete user attempted with no logged in user!');
+      return;
+    }
+    await currentUser?.user?.delete();
+  } on FirebaseAuthException catch (e) {
+    if (e.code == 'requires-recent-login') {
+      ScaffoldMessenger.of(context).hideCurrentSnackBar();
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+            content: Text(
+                'Too long since most recent sign in. Sign in again before deleting your account.')),
+      );
+    }
+  }
+}
+
+Future resetPassword(
+    {required String email, required BuildContext context}) async {
   try {
     await FirebaseAuth.instance.sendPasswordResetEmail(email: email);
   } on FirebaseAuthException catch (e) {
     ScaffoldMessenger.of(context).hideCurrentSnackBar();
     ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(content: Text('Error: ${e.message}')),
+      SnackBar(content: Text('Error: ${e.message!}')),
     );
     return null;
   }
   ScaffoldMessenger.of(context).showSnackBar(
-    const SnackBar(content: Text('Password reset email sent!')),
+    SnackBar(content: Text('Password reset email sent')),
   );
 }
 
 Future sendEmailVerification() async =>
     currentUser?.user?.sendEmailVerification();
 
-String _currentJwtToken = '';
+String? _currentJwtToken = '';
 
 String get currentUserEmail =>
     currentUserDocument?.email ?? currentUser?.user?.email ?? '';
 
-String get currentUserUid =>
-    currentUserDocument?.uid ?? currentUser?.user?.uid ?? '';
+String get currentUserUid => currentUser?.user?.uid ?? '';
 
 String get currentUserDisplayName =>
     currentUserDocument?.displayName ?? currentUser?.user?.displayName ?? '';
@@ -74,23 +97,23 @@ String get currentJwtToken => _currentJwtToken ?? '';
 bool get currentUserEmailVerified {
   // Reloads the user when checking in order to get the most up to date
   // email verified status.
-  if (currentUser?.user != null && !currentUser.user.emailVerified) {
-    currentUser.user
+  if (currentUser?.user != null && !currentUser!.user!.emailVerified) {
+    currentUser!.user!
         .reload()
-        .then((_) => currentUser.user = FirebaseAuth.instance.currentUser);
+        .then((_) => currentUser!.user = FirebaseAuth.instance.currentUser);
   }
   return currentUser?.user?.emailVerified ?? false;
 }
 
 // Set when using phone verification (after phone number is provided).
-String _phoneAuthVerificationCode;
+String? _phoneAuthVerificationCode;
 // Set when using phone sign in in web mode (ignored otherwise).
-ConfirmationResult _webPhoneAuthConfirmationResult;
+ConfirmationResult? _webPhoneAuthConfirmationResult;
 
 Future beginPhoneAuth({
-  BuildContext context,
-  String phoneNumber,
-  VoidCallback onCodeSent,
+  required BuildContext context,
+  required String phoneNumber,
+  required VoidCallback onCodeSent,
 }) async {
   if (kIsWeb) {
     _webPhoneAuthConfirmationResult =
@@ -116,9 +139,9 @@ Future beginPhoneAuth({
       //   MaterialPageRoute(builder: (_) => DestinationPage()),
       // );
     },
-    verificationFailed: (exception) {
+    verificationFailed: (e) {
       ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-        content: Text('Error with phone verification: ${exception.message}'),
+        content: Text('Error: ${e.message!}'),
       ));
     },
     codeSent: (verificationId, _) {
@@ -130,15 +153,15 @@ Future beginPhoneAuth({
 }
 
 Future verifySmsCode({
-  BuildContext context,
-  String smsCode,
+  required BuildContext context,
+  required String smsCode,
 }) async {
   if (kIsWeb) {
     return signInOrCreateAccount(
-        context, () => _webPhoneAuthConfirmationResult.confirm(smsCode));
+        context, () => _webPhoneAuthConfirmationResult!.confirm(smsCode));
   } else {
     final authCredential = PhoneAuthProvider.credential(
-        verificationId: _phoneAuthVerificationCode, smsCode: smsCode);
+        verificationId: _phoneAuthVerificationCode!, smsCode: smsCode);
     return signInOrCreateAccount(
       context,
       () => FirebaseAuth.instance.signInWithCredential(authCredential),
@@ -146,11 +169,11 @@ Future verifySmsCode({
   }
 }
 
-DocumentReference get currentUserReference => currentUser?.user != null
-    ? UsersRecord.collection.doc(currentUser.user.uid)
+DocumentReference? get currentUserReference => currentUser?.user != null
+    ? UsersRecord.collection.doc(currentUser!.user!.uid)
     : null;
 
-UsersRecord currentUserDocument;
+UsersRecord? currentUserDocument;
 final authenticatedUserStream = FirebaseAuth.instance
     .authStateChanges()
     .map<String>((user) {
@@ -160,14 +183,17 @@ final authenticatedUserStream = FirebaseAuth.instance
       }();
       return user?.uid ?? '';
     })
-    .switchMap((uid) => queryUsersRecord(
-        queryBuilder: (user) => user.where('uid', isEqualTo: uid),
-        singleRecord: true))
-    .map((users) => currentUserDocument = users.isNotEmpty ? users.first : null)
+    .switchMap(
+      (uid) => uid.isEmpty
+          ? Stream.value(null)
+          : UsersRecord.getDocument(UsersRecord.collection.doc(uid))
+              .handleError((_) {}),
+    )
+    .map((user) => currentUserDocument = user)
     .asBroadcastStream();
 
 class AuthUserStreamWidget extends StatelessWidget {
-  const AuthUserStreamWidget({Key key, this.child}) : super(key: key);
+  const AuthUserStreamWidget({Key? key, required this.child}) : super(key: key);
 
   final Widget child;
 
