@@ -1,18 +1,17 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:rxdart/rxdart.dart';
 
-class Qonvay2FirebaseUser {
-  Qonvay2FirebaseUser(this.user);
-  User user;
+class QonvayFirebaseUser {
+  QonvayFirebaseUser(this.user);
+  User? user;
   bool get loggedIn => user != null;
 }
 
-Qonvay2FirebaseUser currentUser;
+QonvayFirebaseUser? currentUser;
 bool get loggedIn => currentUser?.loggedIn ?? false;
-Stream<Qonvay2FirebaseUser> qonvay2FirebaseUserStream() => FirebaseAuth.instance
+Stream<QonvayFirebaseUser> qonvayFirebaseUserStream() => FirebaseAuth.instance
     .authStateChanges()
     .debounce((user) => user == null && !loggedIn
         ? TimerStream(true, const Duration(seconds: 1))
         : Stream.value(user))
-    .map<Qonvay2FirebaseUser>(
-        (user) => currentUser = Qonvay2FirebaseUser(user));
+    .map<QonvayFirebaseUser>((user) => currentUser = QonvayFirebaseUser(user));
