@@ -1,350 +1,639 @@
+import '../auth/auth_util.dart';
 import '../backend/backend.dart';
+import '../flutter_flow/flutter_flow_animations.dart';
 import '../flutter_flow/flutter_flow_theme.dart';
 import '../flutter_flow/flutter_flow_util.dart';
 import '../order_tracking/order_tracking_widget.dart';
-import '../set_pickup_location/set_pickup_location_widget.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 class DeliveryOrderListWidget extends StatefulWidget {
   const DeliveryOrderListWidget({
-    Key key,
-    this.orderTracking,
+    Key? key,
+    this.bookOrderDetails,
   }) : super(key: key);
 
-  final DocumentReference orderTracking;
+  final DocumentReference? bookOrderDetails;
 
   @override
   _DeliveryOrderListWidgetState createState() =>
       _DeliveryOrderListWidgetState();
 }
 
-class _DeliveryOrderListWidgetState extends State<DeliveryOrderListWidget> {
+class _DeliveryOrderListWidgetState extends State<DeliveryOrderListWidget>
+    with TickerProviderStateMixin {
+  final animationsMap = {
+    'containerOnPageLoadAnimation1': AnimationInfo(
+      trigger: AnimationTrigger.onPageLoad,
+      duration: 600,
+      hideBeforeAnimating: false,
+      fadeIn: true,
+      initialState: AnimationState(
+        offset: Offset(0, 30),
+        scale: 1,
+        opacity: 0,
+      ),
+      finalState: AnimationState(
+        offset: Offset(0, 0),
+        scale: 1,
+        opacity: 1,
+      ),
+    ),
+    'containerOnPageLoadAnimation2': AnimationInfo(
+      trigger: AnimationTrigger.onPageLoad,
+      duration: 600,
+      hideBeforeAnimating: false,
+      fadeIn: true,
+      initialState: AnimationState(
+        offset: Offset(0, 30),
+        scale: 1,
+        opacity: 0,
+      ),
+      finalState: AnimationState(
+        offset: Offset(0, 0),
+        scale: 1,
+        opacity: 1,
+      ),
+    ),
+  };
   final scaffoldKey = GlobalKey<ScaffoldState>();
 
   @override
+  void initState() {
+    super.initState();
+    startPageLoadAnimations(
+      animationsMap.values
+          .where((anim) => anim.trigger == AnimationTrigger.onPageLoad),
+      this,
+    );
+
+    logFirebaseEvent('screen_view',
+        parameters: {'screen_name': 'deliveryOrderList'});
+  }
+
+  @override
   Widget build(BuildContext context) {
-    return StreamBuilder<List<OrderTrackingRecord>>(
-      stream: queryOrderTrackingRecord(
-        singleRecord: true,
+    return Scaffold(
+      key: scaffoldKey,
+      appBar: AppBar(
+        backgroundColor: FlutterFlowTheme.of(context).background,
+        automaticallyImplyLeading: false,
+        title: Text(
+          'Orders',
+          style: FlutterFlowTheme.of(context).title1,
+        ),
+        actions: [],
+        centerTitle: false,
+        elevation: 0,
       ),
-      builder: (context, snapshot) {
-        // Customize what your widget looks like when it's loading.
-        if (!snapshot.hasData) {
-          return Center(
-            child: SizedBox(
-              width: 40,
-              height: 40,
-              child: SpinKitSquareCircle(
-                color: FlutterFlowTheme.primaryColor,
-                size: 40,
-              ),
-            ),
-          );
-        }
-        List<OrderTrackingRecord> deliveryOrderListOrderTrackingRecordList =
-            snapshot.data;
-        // Return an empty Container when the document does not exist.
-        if (snapshot.data.isEmpty) {
-          return Container();
-        }
-        final deliveryOrderListOrderTrackingRecord =
-            deliveryOrderListOrderTrackingRecordList.isNotEmpty
-                ? deliveryOrderListOrderTrackingRecordList.first
-                : null;
-        return Scaffold(
-          key: scaffoldKey,
-          appBar: AppBar(
-            backgroundColor: FlutterFlowTheme.background,
-            automaticallyImplyLeading: false,
-            leading: InkWell(
-              onTap: () async {
-                Navigator.pop(context);
-              },
-              child: Icon(
-                Icons.chevron_left_rounded,
-                color: FlutterFlowTheme.textColor,
-                size: 32,
-              ),
-            ),
-            title: Text(
-              'Orders',
-              style: FlutterFlowTheme.title3,
-            ),
-            actions: [],
-            centerTitle: false,
-            elevation: 0,
-          ),
-          backgroundColor: FlutterFlowTheme.background,
-          body: SingleChildScrollView(
-            child: Column(
-              mainAxisSize: MainAxisSize.max,
-              children: [
-                Container(
-                  width: MediaQuery.of(context).size.width,
-                  height: MediaQuery.of(context).size.height * 1,
-                  decoration: BoxDecoration(
-                    color: FlutterFlowTheme.background,
+      backgroundColor: FlutterFlowTheme.of(context).background,
+      body: GestureDetector(
+        onTap: () => FocusScope.of(context).unfocus(),
+        child: DefaultTabController(
+          length: 2,
+          initialIndex: 0,
+          child: Column(
+            children: [
+              TabBar(
+                isScrollable: true,
+                labelColor: FlutterFlowTheme.of(context).grayLight,
+                labelStyle: FlutterFlowTheme.of(context).title3,
+                indicatorColor: FlutterFlowTheme.of(context).grayDark,
+                indicatorWeight: 5,
+                tabs: [
+                  Tab(
+                    text: 'All Orders',
                   ),
-                  child: Column(
-                    mainAxisSize: MainAxisSize.max,
-                    children: [
-                      Padding(
-                        padding: EdgeInsetsDirectional.fromSTEB(20, 0, 20, 0),
-                        child: Row(
-                          mainAxisSize: MainAxisSize.max,
-                          children: [
-                            Expanded(
-                              child: Padding(
-                                padding:
-                                    EdgeInsetsDirectional.fromSTEB(0, 12, 0, 0),
-                                child: Text(
-                                  'Delivery Orders',
-                                  style: FlutterFlowTheme.bodyText1,
-                                ),
+                  Tab(
+                    text: 'Delivered',
+                  ),
+                ],
+              ),
+              Expanded(
+                child: TabBarView(
+                  children: [
+                    SingleChildScrollView(
+                      child: Column(
+                        mainAxisSize: MainAxisSize.max,
+                        children: [
+                          Padding(
+                            padding:
+                                EdgeInsetsDirectional.fromSTEB(16, 12, 16, 0),
+                            child: Container(
+                              width: double.infinity,
+                              decoration: BoxDecoration(
+                                color: FlutterFlowTheme.of(context).background,
+                                borderRadius: BorderRadius.circular(12),
                               ),
                             ),
-                          ],
-                        ),
-                      ),
-                      Padding(
-                        padding: EdgeInsetsDirectional.fromSTEB(0, 12, 0, 24),
-                        child: StreamBuilder<List<OrderTrackingRecord>>(
-                          stream: queryOrderTrackingRecord(
-                            queryBuilder: (orderTrackingRecord) =>
-                                orderTrackingRecord
-                                    .where('order_id', isEqualTo: 'Order ID')
-                                    .orderBy('order_timestamp',
-                                        descending: true),
                           ),
-                          builder: (context, snapshot) {
-                            // Customize what your widget looks like when it's loading.
-                            if (!snapshot.hasData) {
-                              return Center(
-                                child: SizedBox(
-                                  width: 40,
-                                  height: 40,
-                                  child: SpinKitSquareCircle(
-                                    color: FlutterFlowTheme.primaryColor,
-                                    size: 40,
-                                  ),
+                          Column(
+                            mainAxisSize: MainAxisSize.max,
+                            children: [
+                              StreamBuilder<List<BookOrderRecord>>(
+                                stream: queryBookOrderRecord(
+                                  queryBuilder: (bookOrderRecord) =>
+                                      bookOrderRecord
+                                          .where('booker_id',
+                                              isEqualTo: currentUserUid)
+                                          .orderBy('order_timestamp',
+                                              descending: true),
                                 ),
-                              );
-                            }
-                            List<OrderTrackingRecord>
-                                transactionListOrderTrackingRecordList =
-                                snapshot.data;
-                            if (transactionListOrderTrackingRecordList
-                                .isEmpty) {
-                              return Center(
-                                child: Image.asset(
-                                  'assets/images/noTransactions.png',
-                                  width: 200,
-                                  height: 400,
-                                  fit: BoxFit.fitWidth,
-                                ),
-                              );
-                            }
-                            return Column(
-                              mainAxisSize: MainAxisSize.max,
-                              children: List.generate(
-                                  transactionListOrderTrackingRecordList.length,
-                                  (transactionListIndex) {
-                                final transactionListOrderTrackingRecord =
-                                    transactionListOrderTrackingRecordList[
-                                        transactionListIndex];
-                                return Padding(
-                                  padding: EdgeInsetsDirectional.fromSTEB(
-                                      0, 0, 0, 8),
-                                  child: InkWell(
-                                    onTap: () async {
-                                      await Navigator.push(
-                                        context,
-                                        MaterialPageRoute(
-                                          builder: (context) =>
-                                              SetPickupLocationWidget(),
+                                builder: (context, snapshot) {
+                                  // Customize what your widget looks like when it's loading.
+                                  if (!snapshot.hasData) {
+                                    return Center(
+                                      child: SizedBox(
+                                        width: 50,
+                                        height: 50,
+                                        child: SpinKitFadingGrid(
+                                          color: FlutterFlowTheme.of(context)
+                                              .primaryColor,
+                                          size: 50,
                                         ),
-                                      );
-                                    },
-                                    child: Container(
-                                      width: MediaQuery.of(context).size.width *
-                                          0.92,
-                                      height: 70,
-                                      decoration: BoxDecoration(
-                                        color: FlutterFlowTheme.darkBackground,
-                                        borderRadius: BorderRadius.circular(8),
                                       ),
-                                      child: InkWell(
-                                        onTap: () async {
-                                          await Navigator.push(
-                                            context,
-                                            MaterialPageRoute(
-                                              builder: (context) =>
-                                                  OrderTrackingWidget(),
-                                            ),
-                                          );
-                                        },
-                                        child: Row(
-                                          mainAxisSize: MainAxisSize.max,
-                                          children: [
-                                            Padding(
-                                              padding: EdgeInsetsDirectional
-                                                  .fromSTEB(8, 0, 0, 0),
-                                              child: Card(
-                                                clipBehavior:
-                                                    Clip.antiAliasWithSaveLayer,
-                                                color:
-                                                    FlutterFlowTheme.background,
-                                                shape: RoundedRectangleBorder(
-                                                  borderRadius:
-                                                      BorderRadius.circular(40),
-                                                ),
-                                                child: Padding(
-                                                  padding: EdgeInsetsDirectional
-                                                      .fromSTEB(8, 8, 8, 8),
-                                                  child: Icon(
-                                                    Icons.speed_sharp,
-                                                    color: FlutterFlowTheme
-                                                        .tertiaryColor,
-                                                    size: 24,
-                                                  ),
+                                    );
+                                  }
+                                  List<BookOrderRecord>
+                                      listViewBookOrderRecordList =
+                                      snapshot.data!;
+                                  if (listViewBookOrderRecordList.isEmpty) {
+                                    return Center(
+                                      child: Image.asset(
+                                        'assets/images/No_Orders.png',
+                                      ),
+                                    );
+                                  }
+                                  return ListView.builder(
+                                    padding: EdgeInsets.zero,
+                                    primary: false,
+                                    shrinkWrap: true,
+                                    scrollDirection: Axis.vertical,
+                                    itemCount:
+                                        listViewBookOrderRecordList.length,
+                                    itemBuilder: (context, listViewIndex) {
+                                      final listViewBookOrderRecord =
+                                          listViewBookOrderRecordList[
+                                              listViewIndex];
+                                      return Padding(
+                                        padding: EdgeInsetsDirectional.fromSTEB(
+                                            16, 4, 16, 8),
+                                        child: InkWell(
+                                          onTap: () async {
+                                            logFirebaseEvent(
+                                                'DELIVERY_ORDER_LIST_Container_yzjeaq9h_O');
+                                            logFirebaseEvent(
+                                                'Container_Navigate-To');
+                                            await Navigator.push(
+                                              context,
+                                              PageTransition(
+                                                type: PageTransitionType.fade,
+                                                duration:
+                                                    Duration(milliseconds: 300),
+                                                reverseDuration:
+                                                    Duration(milliseconds: 300),
+                                                child: OrderTrackingWidget(
+                                                  orderRefParam:
+                                                      listViewBookOrderRecord
+                                                          .reference,
                                                 ),
                                               ),
+                                            );
+                                          },
+                                          child: Container(
+                                            width: double.infinity,
+                                            decoration: BoxDecoration(
+                                              color:
+                                                  FlutterFlowTheme.of(context)
+                                                      .darkBackground,
+                                              borderRadius:
+                                                  BorderRadius.circular(12),
                                             ),
-                                            Expanded(
+                                            child: Padding(
+                                              padding: EdgeInsetsDirectional
+                                                  .fromSTEB(12, 12, 12, 12),
+                                              child: Column(
+                                                mainAxisSize: MainAxisSize.max,
+                                                crossAxisAlignment:
+                                                    CrossAxisAlignment.start,
+                                                children: [
+                                                  Row(
+                                                    mainAxisSize:
+                                                        MainAxisSize.max,
+                                                    mainAxisAlignment:
+                                                        MainAxisAlignment
+                                                            .spaceBetween,
+                                                    children: [
+                                                      Expanded(
+                                                        child: Padding(
+                                                          padding:
+                                                              EdgeInsetsDirectional
+                                                                  .fromSTEB(0,
+                                                                      0, 12, 0),
+                                                          child: Text(
+                                                            listViewBookOrderRecord
+                                                                .receiverName!
+                                                                .maybeHandleOverflow(
+                                                              maxChars: 12,
+                                                              replacement: '…',
+                                                            ),
+                                                            style: FlutterFlowTheme
+                                                                    .of(context)
+                                                                .title3,
+                                                          ),
+                                                        ),
+                                                      ),
+                                                      Container(
+                                                        height: 20,
+                                                        decoration:
+                                                            BoxDecoration(
+                                                          color: FlutterFlowTheme
+                                                                  .of(context)
+                                                              .primaryColor,
+                                                          borderRadius:
+                                                              BorderRadius
+                                                                  .circular(5),
+                                                        ),
+                                                        alignment:
+                                                            AlignmentDirectional(
+                                                                0, 0),
+                                                        child: Padding(
+                                                          padding:
+                                                              EdgeInsetsDirectional
+                                                                  .fromSTEB(2,
+                                                                      0, 2, 0),
+                                                          child: Text(
+                                                            listViewBookOrderRecord
+                                                                .orderStatus!,
+                                                            style: FlutterFlowTheme
+                                                                    .of(context)
+                                                                .bodyText1
+                                                                .override(
+                                                                  fontFamily:
+                                                                      'Lexend Deca',
+                                                                  color: FlutterFlowTheme.of(
+                                                                          context)
+                                                                      .darkBackground,
+                                                                ),
+                                                          ),
+                                                        ),
+                                                      ),
+                                                    ],
+                                                  ),
+                                                  Divider(
+                                                    thickness: 1,
+                                                    color: FlutterFlowTheme.of(
+                                                            context)
+                                                        .grayDark,
+                                                  ),
+                                                  Padding(
+                                                    padding:
+                                                        EdgeInsetsDirectional
+                                                            .fromSTEB(
+                                                                0, 0, 0, 5),
+                                                    child: Row(
+                                                      mainAxisSize:
+                                                          MainAxisSize.max,
+                                                      mainAxisAlignment:
+                                                          MainAxisAlignment
+                                                              .start,
+                                                      children: [
+                                                        Text(
+                                                          listViewBookOrderRecord
+                                                              .orderMileageCost!
+                                                              .toString(),
+                                                          style: FlutterFlowTheme
+                                                                  .of(context)
+                                                              .bodyText1
+                                                              .override(
+                                                                fontFamily:
+                                                                    'Lexend Deca',
+                                                                fontWeight:
+                                                                    FontWeight
+                                                                        .bold,
+                                                              ),
+                                                        ),
+                                                        Padding(
+                                                          padding:
+                                                              EdgeInsetsDirectional
+                                                                  .fromSTEB(1,
+                                                                      0, 0, 0),
+                                                          child: Text(
+                                                            'km',
+                                                            style: FlutterFlowTheme
+                                                                    .of(context)
+                                                                .bodyText1,
+                                                          ),
+                                                        ),
+                                                        Padding(
+                                                          padding:
+                                                              EdgeInsetsDirectional
+                                                                  .fromSTEB(8,
+                                                                      0, 0, 0),
+                                                          child: Text(
+                                                            dateTimeFormat(
+                                                                'yMMMd',
+                                                                listViewBookOrderRecord
+                                                                    .orderTimestamp!),
+                                                            style: FlutterFlowTheme
+                                                                    .of(context)
+                                                                .bodyText2,
+                                                          ),
+                                                        ),
+                                                        if (listViewBookOrderRecord
+                                                                .reportOrderActive ??
+                                                            true)
+                                                          Padding(
+                                                            padding:
+                                                                EdgeInsetsDirectional
+                                                                    .fromSTEB(
+                                                                        8,
+                                                                        0,
+                                                                        0,
+                                                                        0),
+                                                            child: Text(
+                                                              'Under Review',
+                                                              style: FlutterFlowTheme
+                                                                      .of(context)
+                                                                  .bodyText1
+                                                                  .override(
+                                                                    fontFamily:
+                                                                        'Lexend Deca',
+                                                                    color: FlutterFlowTheme.of(
+                                                                            context)
+                                                                        .errorRed,
+                                                                  ),
+                                                            ),
+                                                          ),
+                                                      ],
+                                                    ),
+                                                  ),
+                                                ],
+                                              ),
+                                            ),
+                                          ),
+                                        ).animated([
+                                          animationsMap[
+                                              'containerOnPageLoadAnimation1']!
+                                        ]),
+                                      );
+                                    },
+                                  );
+                                },
+                              ),
+                            ],
+                          ),
+                        ],
+                      ),
+                    ),
+                    SingleChildScrollView(
+                      child: Column(
+                        mainAxisSize: MainAxisSize.max,
+                        children: [
+                          Padding(
+                            padding:
+                                EdgeInsetsDirectional.fromSTEB(16, 12, 16, 0),
+                            child: Container(
+                              width: double.infinity,
+                              decoration: BoxDecoration(
+                                color: FlutterFlowTheme.of(context).background,
+                                borderRadius: BorderRadius.circular(12),
+                              ),
+                            ),
+                          ),
+                          SingleChildScrollView(
+                            child: Column(
+                              mainAxisSize: MainAxisSize.max,
+                              children: [
+                                StreamBuilder<List<BookOrderRecord>>(
+                                  stream: queryBookOrderRecord(
+                                    queryBuilder: (bookOrderRecord) =>
+                                        bookOrderRecord
+                                            .where('booker_id',
+                                                isEqualTo: currentUserUid)
+                                            .where('order_status',
+                                                isEqualTo: 'Item Delivered')
+                                            .orderBy('order_timestamp',
+                                                descending: true),
+                                  ),
+                                  builder: (context, snapshot) {
+                                    // Customize what your widget looks like when it's loading.
+                                    if (!snapshot.hasData) {
+                                      return Center(
+                                        child: SizedBox(
+                                          width: 50,
+                                          height: 50,
+                                          child: SpinKitFadingGrid(
+                                            color: FlutterFlowTheme.of(context)
+                                                .primaryColor,
+                                            size: 50,
+                                          ),
+                                        ),
+                                      );
+                                    }
+                                    List<BookOrderRecord>
+                                        listViewBookOrderRecordList =
+                                        snapshot.data!;
+                                    if (listViewBookOrderRecordList.isEmpty) {
+                                      return Center(
+                                        child: Image.asset(
+                                          'assets/images/No_Orders.png',
+                                        ),
+                                      );
+                                    }
+                                    return ListView.builder(
+                                      padding: EdgeInsets.zero,
+                                      primary: false,
+                                      shrinkWrap: true,
+                                      scrollDirection: Axis.vertical,
+                                      itemCount:
+                                          listViewBookOrderRecordList.length,
+                                      itemBuilder: (context, listViewIndex) {
+                                        final listViewBookOrderRecord =
+                                            listViewBookOrderRecordList[
+                                                listViewIndex];
+                                        return Padding(
+                                          padding:
+                                              EdgeInsetsDirectional.fromSTEB(
+                                                  16, 4, 16, 8),
+                                          child: InkWell(
+                                            onTap: () async {
+                                              logFirebaseEvent(
+                                                  'DELIVERY_ORDER_LIST_Container_u1558272_O');
+                                              logFirebaseEvent(
+                                                  'Container_Navigate-To');
+                                              await Navigator.push(
+                                                context,
+                                                PageTransition(
+                                                  type: PageTransitionType.fade,
+                                                  duration: Duration(
+                                                      milliseconds: 300),
+                                                  reverseDuration: Duration(
+                                                      milliseconds: 300),
+                                                  child: OrderTrackingWidget(
+                                                    orderRefParam:
+                                                        listViewBookOrderRecord
+                                                            .reference,
+                                                  ),
+                                                ),
+                                              );
+                                            },
+                                            child: Container(
+                                              width: double.infinity,
+                                              decoration: BoxDecoration(
+                                                color:
+                                                    FlutterFlowTheme.of(context)
+                                                        .darkBackground,
+                                                borderRadius:
+                                                    BorderRadius.circular(12),
+                                              ),
                                               child: Padding(
                                                 padding: EdgeInsetsDirectional
-                                                    .fromSTEB(12, 0, 0, 0),
+                                                    .fromSTEB(12, 12, 12, 12),
                                                 child: Column(
                                                   mainAxisSize:
                                                       MainAxisSize.max,
-                                                  mainAxisAlignment:
-                                                      MainAxisAlignment.center,
                                                   crossAxisAlignment:
                                                       CrossAxisAlignment.start,
                                                   children: [
-                                                    Text(
-                                                      transactionListOrderTrackingRecord
-                                                          .orderId,
-                                                      style: FlutterFlowTheme
-                                                          .subtitle1
-                                                          .override(
-                                                        fontFamily:
-                                                            'Lexend Deca',
-                                                        color: FlutterFlowTheme
-                                                            .textColor,
-                                                      ),
+                                                    Row(
+                                                      mainAxisSize:
+                                                          MainAxisSize.max,
+                                                      mainAxisAlignment:
+                                                          MainAxisAlignment
+                                                              .spaceBetween,
+                                                      children: [
+                                                        Expanded(
+                                                          child: Padding(
+                                                            padding:
+                                                                EdgeInsetsDirectional
+                                                                    .fromSTEB(
+                                                                        0,
+                                                                        0,
+                                                                        12,
+                                                                        0),
+                                                            child: Text(
+                                                              listViewBookOrderRecord
+                                                                  .receiverName!
+                                                                  .maybeHandleOverflow(
+                                                                maxChars: 12,
+                                                                replacement:
+                                                                    '…',
+                                                              ),
+                                                              style: FlutterFlowTheme
+                                                                      .of(context)
+                                                                  .title3,
+                                                            ),
+                                                          ),
+                                                        ),
+                                                        Icon(
+                                                          Icons.fact_check,
+                                                          color: FlutterFlowTheme
+                                                                  .of(context)
+                                                              .primaryColor,
+                                                          size: 24,
+                                                        ),
+                                                      ],
+                                                    ),
+                                                    Divider(
+                                                      thickness: 1,
+                                                      color:
+                                                          FlutterFlowTheme.of(
+                                                                  context)
+                                                              .grayDark,
                                                     ),
                                                     Padding(
                                                       padding:
                                                           EdgeInsetsDirectional
                                                               .fromSTEB(
-                                                                  0, 4, 0, 0),
-                                                      child: Text(
-                                                        'In Transit',
-                                                        style: FlutterFlowTheme
-                                                            .bodyText1,
+                                                                  0, 0, 0, 5),
+                                                      child: Row(
+                                                        mainAxisSize:
+                                                            MainAxisSize.max,
+                                                        mainAxisAlignment:
+                                                            MainAxisAlignment
+                                                                .start,
+                                                        children: [
+                                                          Text(
+                                                            listViewBookOrderRecord
+                                                                .orderMileageCost!
+                                                                .toString(),
+                                                            style: FlutterFlowTheme
+                                                                    .of(context)
+                                                                .bodyText1
+                                                                .override(
+                                                                  fontFamily:
+                                                                      'Lexend Deca',
+                                                                  fontWeight:
+                                                                      FontWeight
+                                                                          .bold,
+                                                                ),
+                                                          ),
+                                                          Padding(
+                                                            padding:
+                                                                EdgeInsetsDirectional
+                                                                    .fromSTEB(
+                                                                        1,
+                                                                        0,
+                                                                        0,
+                                                                        0),
+                                                            child: Text(
+                                                              'km',
+                                                              style: FlutterFlowTheme
+                                                                      .of(context)
+                                                                  .bodyText1,
+                                                            ),
+                                                          ),
+                                                          Padding(
+                                                            padding:
+                                                                EdgeInsetsDirectional
+                                                                    .fromSTEB(
+                                                                        8,
+                                                                        0,
+                                                                        0,
+                                                                        0),
+                                                            child: Text(
+                                                              dateTimeFormat(
+                                                                  'yMMMd',
+                                                                  listViewBookOrderRecord
+                                                                      .orderTimestamp!),
+                                                              style: FlutterFlowTheme
+                                                                      .of(context)
+                                                                  .bodyText2,
+                                                            ),
+                                                          ),
+                                                        ],
                                                       ),
                                                     ),
                                                   ],
                                                 ),
                                               ),
                                             ),
-                                            Padding(
-                                              padding: EdgeInsetsDirectional
-                                                  .fromSTEB(12, 0, 12, 0),
-                                              child: Column(
-                                                mainAxisSize: MainAxisSize.max,
-                                                mainAxisAlignment:
-                                                    MainAxisAlignment.center,
-                                                crossAxisAlignment:
-                                                    CrossAxisAlignment.end,
-                                                children: [
-                                                  Text(
-                                                    transactionListOrderTrackingRecord
-                                                        .orderMileageCost
-                                                        .toString(),
-                                                    textAlign: TextAlign.end,
-                                                    style: FlutterFlowTheme
-                                                        .subtitle2
-                                                        .override(
-                                                      fontFamily: 'Lexend Deca',
-                                                      color: FlutterFlowTheme
-                                                          .tertiaryColor,
-                                                    ),
-                                                  ),
-                                                  Padding(
-                                                    padding:
-                                                        EdgeInsetsDirectional
-                                                            .fromSTEB(
-                                                                0, 4, 0, 0),
-                                                    child: StreamBuilder<
-                                                        List<
-                                                            OrderTrackingRecord>>(
-                                                      stream:
-                                                          queryOrderTrackingRecord(),
-                                                      builder:
-                                                          (context, snapshot) {
-                                                        // Customize what your widget looks like when it's loading.
-                                                        if (!snapshot.hasData) {
-                                                          return Center(
-                                                            child: SizedBox(
-                                                              width: 40,
-                                                              height: 40,
-                                                              child:
-                                                                  SpinKitSquareCircle(
-                                                                color: FlutterFlowTheme
-                                                                    .primaryColor,
-                                                                size: 40,
-                                                              ),
-                                                            ),
-                                                          );
-                                                        }
-                                                        List<OrderTrackingRecord>
-                                                            orderDateandtimeOrderTrackingRecordList =
-                                                            snapshot.data;
-                                                        return Text(
-                                                          transactionListOrderTrackingRecord
-                                                              .orderTimestamp
-                                                              .toString(),
-                                                          textAlign:
-                                                              TextAlign.end,
-                                                          style:
-                                                              FlutterFlowTheme
-                                                                  .bodyText1
-                                                                  .override(
-                                                            fontFamily:
-                                                                'Lexend Deca',
-                                                            fontSize: 12,
-                                                          ),
-                                                        );
-                                                      },
-                                                    ),
-                                                  ),
-                                                ],
-                                              ),
-                                            ),
-                                          ],
-                                        ),
-                                      ),
-                                    ),
-                                  ),
-                                );
-                              }),
-                            );
-                          },
-                        ),
+                                          ).animated([
+                                            animationsMap[
+                                                'containerOnPageLoadAnimation2']!
+                                          ]),
+                                        );
+                                      },
+                                    );
+                                  },
+                                ),
+                              ],
+                            ),
+                          ),
+                        ],
                       ),
-                    ],
-                  ),
+                    ),
+                  ],
                 ),
-              ],
-            ),
+              ),
+            ],
           ),
-        );
-      },
+        ),
+      ),
     );
   }
 }
