@@ -1,49 +1,88 @@
 import 'package:flutter/material.dart';
 
-class FlutterFlowIconButton extends StatelessWidget {
-  const FlutterFlowIconButton(
-      {Key key,
-      this.borderColor,
-      this.borderRadius,
-      this.borderWidth,
-      this.buttonSize,
-      this.fillColor,
-      this.icon,
-      this.onPressed})
-      : super(key: key);
+class FlutterFlowIconButton extends StatefulWidget {
+  const FlutterFlowIconButton({
+    Key? key,
+    required this.icon,
+    this.borderColor,
+    this.borderRadius,
+    this.borderWidth,
+    this.buttonSize,
+    this.fillColor,
+    this.onPressed,
+    this.showLoadingIndicator = false,
+  }) : super(key: key);
 
-  final double borderRadius;
-  final double buttonSize;
-  final Color fillColor;
-  final Color borderColor;
-  final double borderWidth;
   final Widget icon;
-  final void Function() onPressed;
+  final double? borderRadius;
+  final double? buttonSize;
+  final Color? fillColor;
+  final Color? borderColor;
+  final double? borderWidth;
+  final bool showLoadingIndicator;
+  final Function()? onPressed;
 
   @override
-  Widget build(BuildContext context) => Material(
-        borderRadius:
-            borderRadius != null ? BorderRadius.circular(borderRadius) : null,
-        color: Colors.transparent,
-        clipBehavior: Clip.antiAlias,
-        child: Ink(
-          width: buttonSize,
-          height: buttonSize,
-          decoration: BoxDecoration(
-            color: fillColor,
-            border: Border.all(
-              color: borderColor ?? Colors.transparent,
-              width: borderWidth ?? 0,
-            ),
-            borderRadius: borderRadius != null
-                ? BorderRadius.circular(borderRadius)
-                : null,
+  State<FlutterFlowIconButton> createState() => _FlutterFlowIconButtonState();
+}
+
+class _FlutterFlowIconButtonState extends State<FlutterFlowIconButton> {
+  bool loading = false;
+
+  @override
+  Widget build(BuildContext context) {
+    final icon = widget.icon as Icon;
+    return Material(
+      borderRadius: widget.borderRadius != null
+          ? BorderRadius.circular(widget.borderRadius!)
+          : null,
+      color: Colors.transparent,
+      clipBehavior: Clip.antiAlias,
+      child: Ink(
+        width: widget.buttonSize,
+        height: widget.buttonSize,
+        decoration: BoxDecoration(
+          color: widget.fillColor,
+          border: Border.all(
+            color: widget.borderColor ?? Colors.transparent,
+            width: widget.borderWidth ?? 0,
           ),
-          child: IconButton(
-            icon: icon,
-            onPressed: onPressed,
-            splashRadius: buttonSize,
-          ),
+          borderRadius: widget.borderRadius != null
+              ? BorderRadius.circular(widget.borderRadius!)
+              : null,
         ),
-      );
+        child: (widget.showLoadingIndicator && loading)
+            ? Center(
+                child: Container(
+                  width: icon.size,
+                  height: icon.size,
+                  child: CircularProgressIndicator(
+                    valueColor: AlwaysStoppedAnimation<Color>(
+                      icon.color ?? Colors.white,
+                    ),
+                  ),
+                ),
+              )
+            : IconButton(
+                icon: widget.icon,
+                onPressed: widget.onPressed == null
+                    ? null
+                    : () async {
+                        if (loading) {
+                          return;
+                        }
+                        setState(() => loading = true);
+                        try {
+                          await widget.onPressed!();
+                        } finally {
+                          if (mounted) {
+                            setState(() => loading = false);
+                          }
+                        }
+                      },
+                splashRadius: widget.buttonSize,
+              ),
+      ),
+    );
+  }
 }
