@@ -1,22 +1,23 @@
-import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'auth/firebase_user_provider.dart';
 import 'auth/auth_util.dart';
 
-import '../flutter_flow/flutter_flow_theme.dart';
-import 'package:qonvay_2/login_page/login_page_widget.dart';
 import 'flutter_flow/flutter_flow_theme.dart';
-import 'package:google_fonts/google_fonts.dart';
+import 'flutter_flow/flutter_flow_util.dart';
+import 'flutter_flow/internationalization.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
-import 'main_dashboard/main_dashboard_widget.dart';
-import 'book_delivery/book_delivery_widget.dart';
-import 'm_y_profile_page/m_y_profile_page_widget.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'index.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
+
+  FFAppState(); // Initialize FFAppState
 
   runApp(MyApp());
 }
@@ -24,22 +25,31 @@ void main() async {
 class MyApp extends StatefulWidget {
   // This widget is the root of your application.
   @override
-  _MyAppState createState() => _MyAppState();
+  State<MyApp> createState() => _MyAppState();
+
+  static _MyAppState of(BuildContext context) =>
+      context.findAncestorStateOfType<_MyAppState>()!;
 }
 
 class _MyAppState extends State<MyApp> {
-  Stream<Qonvay2FirebaseUser> userStream;
-  Qonvay2FirebaseUser initialUser;
+  Locale? _locale;
+  ThemeMode _themeMode = ThemeMode.system;
+
+  late Stream<QonvayFirebaseUser> userStream;
+  QonvayFirebaseUser? initialUser;
   bool displaySplashImage = true;
+
   final authUserSub = authenticatedUserStream.listen((_) {});
 
   @override
   void initState() {
     super.initState();
-    userStream = qonvay2FirebaseUserStream()
+    userStream = qonvayFirebaseUserStream()
       ..listen((user) => initialUser ?? setState(() => initialUser = user));
     Future.delayed(
-        Duration(seconds: 1), () => setState(() => displaySplashImage = false));
+      Duration(seconds: 1),
+      () => setState(() => displaySplashImage = false),
+    );
   }
 
   @override
@@ -49,31 +59,42 @@ class _MyAppState extends State<MyApp> {
     super.dispose();
   }
 
+  void setLocale(String language) =>
+      setState(() => _locale = createLocale(language));
+  void setThemeMode(ThemeMode mode) => setState(() {
+        _themeMode = mode;
+      });
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Qonvay 2',
+      title: 'Qonvay',
       localizationsDelegates: [
+        FFLocalizationsDelegate(),
         GlobalMaterialLocalizations.delegate,
         GlobalWidgetsLocalizations.delegate,
         GlobalCupertinoLocalizations.delegate,
       ],
-      supportedLocales: const [Locale('en', '')],
-      theme: ThemeData(primarySwatch: Colors.blue),
+      locale: _locale,
+      supportedLocales: const [
+        Locale('en'),
+      ],
+      theme: ThemeData(brightness: Brightness.light),
+      themeMode: _themeMode,
       home: initialUser == null || displaySplashImage
           ? Container(
-              color: FlutterFlowTheme.background,
+              color: FlutterFlowTheme.of(context).darkBackground,
               child: Center(
                 child: Builder(
                   builder: (context) => Image.asset(
                     'assets/images/qonvayappsplash.svg',
                     width: MediaQuery.of(context).size.width,
-                    fit: BoxFit.cover,
+                    fit: BoxFit.fitWidth,
                   ),
                 ),
               ),
             )
-          : currentUser.loggedIn
+          : currentUser!.loggedIn
               ? NavBarPage()
               : LoginPageWidget(),
     );
@@ -81,9 +102,10 @@ class _MyAppState extends State<MyApp> {
 }
 
 class NavBarPage extends StatefulWidget {
-  NavBarPage({Key key, this.initialPage}) : super(key: key);
+  NavBarPage({Key? key, this.initialPage, this.page}) : super(key: key);
 
-  final String initialPage;
+  final String? initialPage;
+  final Widget? page;
 
   @override
   _NavBarPageState createState() => _NavBarPageState();
@@ -91,12 +113,14 @@ class NavBarPage extends StatefulWidget {
 
 /// This is the private State class that goes with NavBarPage.
 class _NavBarPageState extends State<NavBarPage> {
-  String _currentPage = 'mainDashboard';
+  String _currentPageName = 'mainDashboard';
+  late Widget? _currentPage;
 
   @override
   void initState() {
     super.initState();
-    _currentPage = widget.initialPage ?? _currentPage;
+    _currentPageName = widget.initialPage ?? _currentPageName;
+    _currentPage = widget.page;
   }
 
   @override
@@ -104,21 +128,26 @@ class _NavBarPageState extends State<NavBarPage> {
     final tabs = {
       'mainDashboard': MainDashboardWidget(),
       'bookDelivery': BookDeliveryWidget(),
+      'deliveryOrderList': DeliveryOrderListWidget(),
+      'paymentPage': PaymentPageWidget(),
       'MY_profilePage': MYProfilePageWidget(),
     };
-    final currentIndex = tabs.keys.toList().indexOf(_currentPage);
+    final currentIndex = tabs.keys.toList().indexOf(_currentPageName);
     return Scaffold(
-      body: tabs[_currentPage],
+      body: _currentPage ?? tabs[_currentPageName],
       bottomNavigationBar: BottomNavigationBar(
         currentIndex: currentIndex,
-        onTap: (i) => setState(() => _currentPage = tabs.keys.toList()[i]),
-        backgroundColor: FlutterFlowTheme.darkBackground,
-        selectedItemColor: FlutterFlowTheme.primaryColor,
-        unselectedItemColor: FlutterFlowTheme.grayLight,
-        showSelectedLabels: true,
+        onTap: (i) => setState(() {
+          _currentPage = null;
+          _currentPageName = tabs.keys.toList()[i];
+        }),
+        backgroundColor: FlutterFlowTheme.of(context).darkBackground,
+        selectedItemColor: FlutterFlowTheme.of(context).primaryColor,
+        unselectedItemColor: FlutterFlowTheme.of(context).grayLight,
+        showSelectedLabels: false,
         showUnselectedLabels: false,
         type: BottomNavigationBarType.fixed,
-        items: const <BottomNavigationBarItem>[
+        items: <BottomNavigationBarItem>[
           BottomNavigationBarItem(
             icon: Icon(
               Icons.home,
@@ -128,19 +157,35 @@ class _NavBarPageState extends State<NavBarPage> {
               Icons.home,
               size: 20,
             ),
-            label: '•',
+            label: 'Dashboard',
             tooltip: '',
           ),
           BottomNavigationBarItem(
             icon: Icon(
-              Icons.delivery_dining,
+              Icons.add_box,
               size: 24,
             ),
             activeIcon: Icon(
-              Icons.delivery_dining,
+              Icons.add_box,
               size: 24,
             ),
-            label: '•',
+            label: 'Book Delivery',
+            tooltip: '',
+          ),
+          BottomNavigationBarItem(
+            icon: FaIcon(
+              FontAwesomeIcons.clipboardList,
+              size: 24,
+            ),
+            label: 'Your Bookings',
+            tooltip: '',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(
+              Icons.payment,
+              size: 24,
+            ),
+            label: 'Payment',
             tooltip: '',
           ),
           BottomNavigationBarItem(
@@ -152,7 +197,7 @@ class _NavBarPageState extends State<NavBarPage> {
               Icons.account_circle_rounded,
               size: 24,
             ),
-            label: '•',
+            label: 'Profile',
             tooltip: '',
           )
         ],
