@@ -1,18 +1,21 @@
 import '../auth/auth_util.dart';
 import '../backend/backend.dart';
-import '../components/add_card_widget.dart';
 import '../flutter_flow/flutter_flow_animations.dart';
+import '../flutter_flow/flutter_flow_credit_card_form.dart';
 import '../flutter_flow/flutter_flow_icon_button.dart';
 import '../flutter_flow/flutter_flow_theme.dart';
 import '../flutter_flow/flutter_flow_util.dart';
 import '../flutter_flow/flutter_flow_widgets.dart';
+import '../main.dart';
+import '../flutter_flow/custom_functions.dart' as functions;
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:easy_debounce/easy_debounce.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 class TopupMileageWidget extends StatefulWidget {
-  const TopupMileageWidget({Key key}) : super(key: key);
+  const TopupMileageWidget({Key? key}) : super(key: key);
 
   @override
   _TopupMileageWidgetState createState() => _TopupMileageWidgetState();
@@ -20,14 +23,17 @@ class TopupMileageWidget extends StatefulWidget {
 
 class _TopupMileageWidgetState extends State<TopupMileageWidget>
     with TickerProviderStateMixin {
-  TextEditingController amountNGNmileageController;
-  TextEditingController deliveryMileageDisplayController;
+  TextEditingController? amountNGNmileageController;
+
+  final creditCardFormKey = GlobalKey<FormState>();
+  CreditCardModel creditCardInfo = emptyCreditCard();
   final formKey = GlobalKey<FormState>();
   final scaffoldKey = GlobalKey<ScaffoldState>();
   final animationsMap = {
-    'textFieldOnPageLoadAnimation1': AnimationInfo(
+    'textFieldOnPageLoadAnimation': AnimationInfo(
       trigger: AnimationTrigger.onPageLoad,
       duration: 600,
+      hideBeforeAnimating: false,
       fadeIn: true,
       initialState: AnimationState(
         offset: Offset(0, 40),
@@ -35,35 +41,6 @@ class _TopupMileageWidgetState extends State<TopupMileageWidget>
       ),
       finalState: AnimationState(
         offset: Offset(0, 0),
-        opacity: 1,
-      ),
-    ),
-    'textFieldOnPageLoadAnimation2': AnimationInfo(
-      trigger: AnimationTrigger.onPageLoad,
-      duration: 600,
-      delay: 170,
-      fadeIn: true,
-      initialState: AnimationState(
-        offset: Offset(0, 80),
-        opacity: 0,
-      ),
-      finalState: AnimationState(
-        offset: Offset(0, 0),
-        opacity: 1,
-      ),
-    ),
-    'textOnPageLoadAnimation': AnimationInfo(
-      trigger: AnimationTrigger.onPageLoad,
-      duration: 600,
-      fadeIn: true,
-      initialState: AnimationState(
-        offset: Offset(0, 0),
-        scale: 1,
-        opacity: 0,
-      ),
-      finalState: AnimationState(
-        offset: Offset(0, 0),
-        scale: 1,
         opacity: 1,
       ),
     ),
@@ -79,279 +56,330 @@ class _TopupMileageWidgetState extends State<TopupMileageWidget>
     );
 
     amountNGNmileageController = TextEditingController();
-    deliveryMileageDisplayController = TextEditingController();
+    logFirebaseEvent('screen_view',
+        parameters: {'screen_name': 'topupMileage'});
   }
 
   @override
   Widget build(BuildContext context) {
-    return Form(
-      key: formKey,
-      child: Scaffold(
-        key: scaffoldKey,
-        backgroundColor: FlutterFlowTheme.darkBackground,
-        body: Column(
-          mainAxisSize: MainAxisSize.max,
-          children: [
-            Material(
-              color: Colors.transparent,
-              elevation: 3,
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.only(
-                  bottomLeft: Radius.circular(16),
-                  bottomRight: Radius.circular(16),
-                  topLeft: Radius.circular(0),
-                  topRight: Radius.circular(0),
-                ),
+    return StreamBuilder<List<UsersRecord>>(
+      stream: queryUsersRecord(
+        singleRecord: true,
+      ),
+      builder: (context, snapshot) {
+        // Customize what your widget looks like when it's loading.
+        if (!snapshot.hasData) {
+          return Center(
+            child: SizedBox(
+              width: 50,
+              height: 50,
+              child: SpinKitFadingGrid(
+                color: FlutterFlowTheme.of(context).primaryColor,
+                size: 50,
               ),
-              child: Container(
-                width: MediaQuery.of(context).size.width,
-                height: MediaQuery.of(context).size.height * 0.8,
-                decoration: BoxDecoration(
-                  color: FlutterFlowTheme.darkBackground,
-                  borderRadius: BorderRadius.only(
-                    bottomLeft: Radius.circular(16),
-                    bottomRight: Radius.circular(16),
-                    topLeft: Radius.circular(0),
-                    topRight: Radius.circular(0),
+            ),
+          );
+        }
+        List<UsersRecord> topupMileageUsersRecordList = snapshot.data!;
+        // Return an empty Container when the document does not exist.
+        if (snapshot.data!.isEmpty) {
+          return Container();
+        }
+        final topupMileageUsersRecord = topupMileageUsersRecordList.isNotEmpty
+            ? topupMileageUsersRecordList.first
+            : null;
+        return Scaffold(
+          key: scaffoldKey,
+          backgroundColor: FlutterFlowTheme.of(context).darkBackground,
+          body: Form(
+            key: formKey,
+            autovalidateMode: AutovalidateMode.disabled,
+            child: Column(
+              mainAxisSize: MainAxisSize.max,
+              children: [
+                Material(
+                  color: Colors.transparent,
+                  elevation: 3,
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(0),
                   ),
-                ),
-                child: Padding(
-                  padding: EdgeInsetsDirectional.fromSTEB(20, 44, 20, 20),
-                  child: Column(
-                    mainAxisSize: MainAxisSize.max,
-                    children: [
-                      Row(
+                  child: Container(
+                    width: MediaQuery.of(context).size.width,
+                    height: MediaQuery.of(context).size.height * 1,
+                    decoration: BoxDecoration(
+                      color: FlutterFlowTheme.of(context).background,
+                      borderRadius: BorderRadius.circular(0),
+                    ),
+                    child: Padding(
+                      padding: EdgeInsetsDirectional.fromSTEB(20, 44, 20, 0),
+                      child: Column(
                         mainAxisSize: MainAxisSize.max,
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
-                          Text(
-                            'Top Up Mileage',
-                            style: FlutterFlowTheme.title1,
-                          ),
-                          Card(
-                            clipBehavior: Clip.antiAliasWithSaveLayer,
-                            color: FlutterFlowTheme.background,
-                            shape: RoundedRectangleBorder(
-                              borderRadius: BorderRadius.circular(30),
-                            ),
-                            child: FlutterFlowIconButton(
-                              borderColor: Colors.transparent,
-                              borderRadius: 30,
-                              buttonSize: 48,
-                              icon: Icon(
-                                Icons.close_rounded,
-                                color: FlutterFlowTheme.textColor,
-                                size: 30,
+                          Row(
+                            mainAxisSize: MainAxisSize.max,
+                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                            children: [
+                              Text(
+                                'Top Up Mileage',
+                                style: FlutterFlowTheme.of(context).title1,
                               ),
+                              Card(
+                                clipBehavior: Clip.antiAliasWithSaveLayer,
+                                color: FlutterFlowTheme.of(context).background,
+                                shape: RoundedRectangleBorder(
+                                  borderRadius: BorderRadius.circular(30),
+                                ),
+                                child: FlutterFlowIconButton(
+                                  borderColor: Colors.transparent,
+                                  borderRadius: 30,
+                                  buttonSize: 48,
+                                  icon: Icon(
+                                    Icons.close_rounded,
+                                    color:
+                                        FlutterFlowTheme.of(context).textColor,
+                                    size: 30,
+                                  ),
+                                  onPressed: () async {
+                                    logFirebaseEvent(
+                                        'TOPUP_MILEAGE_close_rounded_ICN_ON_TAP');
+                                    logFirebaseEvent(
+                                        'IconButton_Update-Local-State');
+                                    setState(
+                                        () => FFAppState().ngnToKmCheck = 0.0);
+                                    logFirebaseEvent(
+                                        'IconButton_Navigate-Back');
+                                    Navigator.pop(context);
+                                  },
+                                ),
+                              ),
+                            ],
+                          ),
+                          Container(
+                            width: MediaQuery.of(context).size.width * 0.8,
+                            height: 100,
+                            constraints: BoxConstraints(
+                              maxWidth: MediaQuery.of(context).size.width * 0.8,
+                            ),
+                            decoration: BoxDecoration(),
+                            child: Padding(
+                              padding:
+                                  EdgeInsetsDirectional.fromSTEB(0, 16, 0, 0),
+                              child: TextFormField(
+                                controller: amountNGNmileageController,
+                                onChanged: (_) => EasyDebounce.debounce(
+                                  'amountNGNmileageController',
+                                  Duration(milliseconds: 500),
+                                  () async {
+                                    logFirebaseEvent(
+                                        'TOPUP_MILEAGE_amountNGNmileage_ON_TEXTFI');
+                                    logFirebaseEvent(
+                                        'amountNGNmileage_Update-Local-State');
+                                    setState(() => FFAppState().ngnToKmCheck =
+                                        functions.ngnToKm(double.parse(
+                                            amountNGNmileageController!.text)));
+                                  },
+                                ),
+                                obscureText: false,
+                                decoration: InputDecoration(
+                                  labelStyle: FlutterFlowTheme.of(context)
+                                      .title1
+                                      .override(
+                                        fontFamily: 'Lexend Deca',
+                                        color: FlutterFlowTheme.of(context)
+                                            .grayLight,
+                                        fontWeight: FontWeight.w300,
+                                      ),
+                                  hintText: 'Amount (NGN)',
+                                  hintStyle: FlutterFlowTheme.of(context)
+                                      .title1
+                                      .override(
+                                        fontFamily: 'Lexend Deca',
+                                        color: FlutterFlowTheme.of(context)
+                                            .grayLight,
+                                        fontWeight: FontWeight.w300,
+                                      ),
+                                  enabledBorder: UnderlineInputBorder(
+                                    borderSide: BorderSide(
+                                      color:
+                                          FlutterFlowTheme.of(context).grayDark,
+                                      width: 2,
+                                    ),
+                                    borderRadius: BorderRadius.circular(8),
+                                  ),
+                                  focusedBorder: UnderlineInputBorder(
+                                    borderSide: BorderSide(
+                                      color:
+                                          FlutterFlowTheme.of(context).grayDark,
+                                      width: 2,
+                                    ),
+                                    borderRadius: BorderRadius.circular(8),
+                                  ),
+                                  errorBorder: UnderlineInputBorder(
+                                    borderSide: BorderSide(
+                                      color: Color(0x00000000),
+                                      width: 2,
+                                    ),
+                                    borderRadius: BorderRadius.circular(8),
+                                  ),
+                                  focusedErrorBorder: UnderlineInputBorder(
+                                    borderSide: BorderSide(
+                                      color: Color(0x00000000),
+                                      width: 2,
+                                    ),
+                                    borderRadius: BorderRadius.circular(8),
+                                  ),
+                                  contentPadding:
+                                      EdgeInsetsDirectional.fromSTEB(
+                                          20, 24, 24, 24),
+                                ),
+                                style: FlutterFlowTheme.of(context).title1,
+                                textAlign: TextAlign.center,
+                                keyboardType:
+                                    const TextInputType.numberWithOptions(
+                                        signed: true, decimal: true),
+                              ).animated([
+                                animationsMap['textFieldOnPageLoadAnimation']!
+                              ]),
+                            ),
+                          ),
+                          Padding(
+                            padding:
+                                EdgeInsetsDirectional.fromSTEB(0, 5, 0, 10),
+                            child: Container(
+                              width: MediaQuery.of(context).size.width * 0.8,
+                              height: 50,
+                              decoration: BoxDecoration(
+                                color: FlutterFlowTheme.of(context)
+                                    .secondaryBackground,
+                                borderRadius: BorderRadius.circular(10),
+                                border: Border.all(
+                                  color:
+                                      FlutterFlowTheme.of(context).primaryColor,
+                                ),
+                              ),
+                              child: Row(
+                                mainAxisSize: MainAxisSize.max,
+                                mainAxisAlignment: MainAxisAlignment.center,
+                                children: [
+                                  Text(
+                                    FFAppState().ngnToKmCheck.toString(),
+                                    style: FlutterFlowTheme.of(context).title3,
+                                  ),
+                                  Padding(
+                                    padding: EdgeInsetsDirectional.fromSTEB(
+                                        2, 0, 0, 0),
+                                    child: Text(
+                                      'km',
+                                      style: FlutterFlowTheme.of(context)
+                                          .bodyText1,
+                                    ),
+                                  ),
+                                ],
+                              ),
+                            ),
+                          ),
+                          Divider(
+                            thickness: 1,
+                            color: FlutterFlowTheme.of(context).grayDark,
+                          ),
+                          Padding(
+                            padding:
+                                EdgeInsetsDirectional.fromSTEB(12, 0, 12, 0),
+                            child: FlutterFlowCreditCardForm(
+                              formKey: creditCardFormKey,
+                              creditCardModel: creditCardInfo,
+                              obscureNumber: true,
+                              obscureCvv: true,
+                              spacing: 10,
+                              textStyle: FlutterFlowTheme.of(context).title3,
+                              inputDecoration: InputDecoration(
+                                enabledBorder: OutlineInputBorder(
+                                  borderSide: BorderSide(
+                                    color:
+                                        FlutterFlowTheme.of(context).grayDark,
+                                    width: 3,
+                                  ),
+                                  borderRadius: BorderRadius.circular(4),
+                                ),
+                                focusedBorder: OutlineInputBorder(
+                                  borderSide: BorderSide(
+                                    color:
+                                        FlutterFlowTheme.of(context).grayDark,
+                                    width: 3,
+                                  ),
+                                  borderRadius: BorderRadius.circular(4),
+                                ),
+                              ),
+                            ),
+                          ),
+                          Padding(
+                            padding:
+                                EdgeInsetsDirectional.fromSTEB(20, 40, 20, 20),
+                            child: FFButtonWidget(
                               onPressed: () async {
-                                Navigator.pop(context);
+                                logFirebaseEvent(
+                                    'TOPUP_MILEAGE_PAGE_TOP_UP_BTN_ON_TAP');
+                                logFirebaseEvent('Button_Backend-Call');
+
+                                final usersUpdateData = {
+                                  'mileage_balance': FieldValue.increment(
+                                      functions.ngnToKm(double.parse(
+                                          amountNGNmileageController!.text))),
+                                };
+                                await currentUserReference!
+                                    .update(usersUpdateData);
+                                logFirebaseEvent('Button_Update-Local-State');
+                                setState(() => FFAppState().ngnToKmCheck = 0.0);
+                                logFirebaseEvent('Button_Navigate-To');
+                                await Navigator.push(
+                                  context,
+                                  PageTransition(
+                                    type: PageTransitionType.fade,
+                                    duration: Duration(milliseconds: 300),
+                                    reverseDuration:
+                                        Duration(milliseconds: 300),
+                                    child: NavBarPage(
+                                        initialPage: 'mainDashboard'),
+                                  ),
+                                );
                               },
+                              text: 'Top Up',
+                              icon: Icon(
+                                Icons.payment,
+                                color:
+                                    FlutterFlowTheme.of(context).darkBackground,
+                                size: 24,
+                              ),
+                              options: FFButtonOptions(
+                                width: 300,
+                                height: 80,
+                                color:
+                                    FlutterFlowTheme.of(context).primaryColor,
+                                textStyle: FlutterFlowTheme.of(context)
+                                    .title1
+                                    .override(
+                                      fontFamily: 'Lexend Deca',
+                                      color: FlutterFlowTheme.of(context)
+                                          .darkBackground,
+                                    ),
+                                borderSide: BorderSide(
+                                  color: Colors.transparent,
+                                  width: 1,
+                                ),
+                                borderRadius: BorderRadius.circular(12),
+                              ),
                             ),
                           ),
                         ],
                       ),
-                      Container(
-                        width: MediaQuery.of(context).size.width * 0.8,
-                        height: 100,
-                        constraints: BoxConstraints(
-                          maxWidth: MediaQuery.of(context).size.width * 0.8,
-                        ),
-                        decoration: BoxDecoration(),
-                        child: Padding(
-                          padding: EdgeInsetsDirectional.fromSTEB(0, 16, 0, 0),
-                          child: TextFormField(
-                            controller: amountNGNmileageController,
-                            obscureText: false,
-                            decoration: InputDecoration(
-                              hintText: 'Amount (NGN)',
-                              hintStyle: FlutterFlowTheme.title1.override(
-                                fontFamily: 'Lexend Deca',
-                                color: FlutterFlowTheme.grayLight,
-                                fontWeight: FontWeight.w300,
-                              ),
-                              enabledBorder: UnderlineInputBorder(
-                                borderSide: BorderSide(
-                                  color: FlutterFlowTheme.background,
-                                  width: 2,
-                                ),
-                                borderRadius: BorderRadius.circular(8),
-                              ),
-                              focusedBorder: UnderlineInputBorder(
-                                borderSide: BorderSide(
-                                  color: FlutterFlowTheme.background,
-                                  width: 2,
-                                ),
-                                borderRadius: BorderRadius.circular(8),
-                              ),
-                              contentPadding: EdgeInsetsDirectional.fromSTEB(
-                                  20, 24, 24, 24),
-                            ),
-                            style: FlutterFlowTheme.title1,
-                            textAlign: TextAlign.center,
-                            keyboardType: TextInputType.number,
-                            validator: (val) {
-                              if (val.isEmpty) {
-                                return 'Please enter an amount';
-                              }
-
-                              return null;
-                            },
-                          ).animated(
-                              [animationsMap['textFieldOnPageLoadAnimation1']]),
-                        ),
-                      ),
-                      Padding(
-                        padding: EdgeInsetsDirectional.fromSTEB(0, 16, 0, 0),
-                        child: TextFormField(
-                          controller: deliveryMileageDisplayController,
-                          obscureText: false,
-                          decoration: InputDecoration(
-                            labelText: 'Delivery Mileage',
-                            labelStyle: FlutterFlowTheme.bodyText1,
-                            enabledBorder: OutlineInputBorder(
-                              borderSide: BorderSide(
-                                color: FlutterFlowTheme.background,
-                                width: 2,
-                              ),
-                              borderRadius: BorderRadius.circular(8),
-                            ),
-                            focusedBorder: OutlineInputBorder(
-                              borderSide: BorderSide(
-                                color: FlutterFlowTheme.background,
-                                width: 2,
-                              ),
-                              borderRadius: BorderRadius.circular(8),
-                            ),
-                            contentPadding:
-                                EdgeInsetsDirectional.fromSTEB(20, 24, 24, 24),
-                          ),
-                          style: FlutterFlowTheme.bodyText2,
-                        ).animated(
-                            [animationsMap['textFieldOnPageLoadAnimation2']]),
-                      ),
-                      Align(
-                        alignment: AlignmentDirectional(0, 0),
-                        child: Padding(
-                          padding: EdgeInsetsDirectional.fromSTEB(0, 10, 0, 0),
-                          child: InkWell(
-                            onTap: () async {
-                              await showModalBottomSheet(
-                                isScrollControlled: true,
-                                context: context,
-                                builder: (context) {
-                                  return Padding(
-                                    padding: MediaQuery.of(context).viewInsets,
-                                    child: AddCardWidget(),
-                                  );
-                                },
-                              );
-                            },
-                            child: Text(
-                              'Add Card',
-                              textAlign: TextAlign.center,
-                              style: FlutterFlowTheme.bodyText1.override(
-                                fontFamily: 'Lexend Deca',
-                                color: FlutterFlowTheme.secondaryColor,
-                                fontSize: 18,
-                              ),
-                            ),
-                          ).animated(
-                              [animationsMap['textOnPageLoadAnimation']]),
-                        ),
-                      ),
-                    ],
+                    ),
                   ),
                 ),
-              ),
+              ],
             ),
-            Padding(
-              padding: EdgeInsetsDirectional.fromSTEB(0, 8, 0, 0),
-              child: Row(
-                mainAxisSize: MainAxisSize.max,
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Row(
-                    mainAxisSize: MainAxisSize.max,
-                    children: [
-                      StreamBuilder<List<AccountRecord>>(
-                        stream: queryAccountRecord(
-                          singleRecord: true,
-                        ),
-                        builder: (context, snapshot) {
-                          // Customize what your widget looks like when it's loading.
-                          if (!snapshot.hasData) {
-                            return Center(
-                              child: SizedBox(
-                                width: 40,
-                                height: 40,
-                                child: SpinKitSquareCircle(
-                                  color: FlutterFlowTheme.primaryColor,
-                                  size: 40,
-                                ),
-                              ),
-                            );
-                          }
-                          List<AccountRecord> buttonAccountRecordList =
-                              snapshot.data;
-                          // Return an empty Container when the document does not exist.
-                          if (snapshot.data.isEmpty) {
-                            return Container();
-                          }
-                          final buttonAccountRecord =
-                              buttonAccountRecordList.isNotEmpty
-                                  ? buttonAccountRecordList.first
-                                  : null;
-                          return FFButtonWidget(
-                            onPressed: () async {
-                              if (!formKey.currentState.validate()) {
-                                return;
-                              }
-                              final transactionsCreateData =
-                                  createTransactionsRecordData(
-                                amount: double.parse(
-                                    amountNGNmileageController.text),
-                                mileagePurchased: double.parse(
-                                    deliveryMileageDisplayController.text),
-                              );
-                              await TransactionsRecord.collection
-                                  .doc()
-                                  .set(transactionsCreateData);
-                              Navigator.pop(context);
-                            },
-                            text: 'Top Up',
-                            options: FFButtonOptions(
-                              width: 300,
-                              height: 70,
-                              color: FlutterFlowTheme.tertiaryColor,
-                              textStyle: FlutterFlowTheme.title1.override(
-                                fontFamily: 'Lexend Deca',
-                                color: FlutterFlowTheme.background,
-                              ),
-                              elevation: 0,
-                              borderSide: BorderSide(
-                                color: Colors.transparent,
-                                width: 1,
-                              ),
-                              borderRadius: 12,
-                            ),
-                          );
-                        },
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-            Text(
-              'Tap above to complete request',
-              style: FlutterFlowTheme.bodyText1.override(
-                fontFamily: 'Lexend Deca',
-                color: Color(0x43000000),
-              ),
-            ),
-          ],
-        ),
-      ),
+          ),
+        );
+      },
     );
   }
 }
