@@ -11,73 +11,90 @@ abstract class BookOrderRecord
   static Serializer<BookOrderRecord> get serializer =>
       _$bookOrderRecordSerializer;
 
-  @nullable
   @BuiltValueField(wireName: 'sender_name')
-  String get senderName;
+  String? get senderName;
 
-  @nullable
   @BuiltValueField(wireName: 'sender_email')
-  String get senderEmail;
+  String? get senderEmail;
 
-  @nullable
   @BuiltValueField(wireName: 'sender_phone')
-  int get senderPhone;
+  int? get senderPhone;
 
-  @nullable
   @BuiltValueField(wireName: 'sender_address')
-  String get senderAddress;
+  String? get senderAddress;
 
-  @nullable
   @BuiltValueField(wireName: 'sender_location')
-  LatLng get senderLocation;
+  LatLng? get senderLocation;
 
-  @nullable
   @BuiltValueField(wireName: 'sender_landmark')
-  String get senderLandmark;
+  String? get senderLandmark;
 
-  @nullable
-  @BuiltValueField(wireName: 'sender_direction')
-  String get senderDirection;
-
-  @nullable
   @BuiltValueField(wireName: 'receiver_name')
-  String get receiverName;
+  String? get receiverName;
 
-  @nullable
   @BuiltValueField(wireName: 'receiver_email')
-  String get receiverEmail;
+  String? get receiverEmail;
 
-  @nullable
   @BuiltValueField(wireName: 'receiver_phone')
-  int get receiverPhone;
+  int? get receiverPhone;
 
-  @nullable
   @BuiltValueField(wireName: 'receiver_address')
-  String get receiverAddress;
+  String? get receiverAddress;
 
-  @nullable
   @BuiltValueField(wireName: 'receiver_location')
-  LatLng get receiverLocation;
+  LatLng? get receiverLocation;
 
-  @nullable
   @BuiltValueField(wireName: 'receiver_landmark')
-  String get receiverLandmark;
+  String? get receiverLandmark;
 
-  @nullable
-  @BuiltValueField(wireName: 'receiver_direction')
-  String get receiverDirection;
-
-  @nullable
   @BuiltValueField(wireName: 'order_timestamp')
-  DateTime get orderTimestamp;
+  DateTime? get orderTimestamp;
 
-  @nullable
   @BuiltValueField(wireName: 'order_mileage_cost')
-  double get orderMileageCost;
+  double? get orderMileageCost;
 
-  @nullable
+  @BuiltValueField(wireName: 'food_category')
+  bool? get foodCategory;
+
+  @BuiltValueField(wireName: 'clothing_category')
+  bool? get clothingCategory;
+
+  @BuiltValueField(wireName: 'shoes_category')
+  bool? get shoesCategory;
+
+  @BuiltValueField(wireName: 'electronics_category')
+  bool? get electronicsCategory;
+
+  @BuiltValueField(wireName: 'jewaccess_category')
+  bool? get jewaccessCategory;
+
+  @BuiltValueField(wireName: 'documents_category')
+  bool? get documentsCategory;
+
+  @BuiltValueField(wireName: 'other_category')
+  bool? get otherCategory;
+
+  @BuiltValueField(wireName: 'order_id')
+  String? get orderId;
+
+  @BuiltValueField(wireName: 'booker_id')
+  String? get bookerId;
+
+  @BuiltValueField(wireName: 'order_status')
+  String? get orderStatus;
+
+  @BuiltValueField(wireName: 'report_order_status')
+  String? get reportOrderStatus;
+
+  @BuiltValueField(wireName: 'report_order_description')
+  String? get reportOrderDescription;
+
+  @BuiltValueField(wireName: 'report_order_active')
+  bool? get reportOrderActive;
+
   @BuiltValueField(wireName: kDocumentReferenceField)
-  DocumentReference get reference;
+  DocumentReference? get ffRef;
+  DocumentReference get reference => ffRef!;
 
   static void _initializeBuilder(BookOrderRecordBuilder builder) => builder
     ..senderName = ''
@@ -85,25 +102,36 @@ abstract class BookOrderRecord
     ..senderPhone = 0
     ..senderAddress = ''
     ..senderLandmark = ''
-    ..senderDirection = ''
     ..receiverName = ''
     ..receiverEmail = ''
     ..receiverPhone = 0
     ..receiverAddress = ''
     ..receiverLandmark = ''
-    ..receiverDirection = ''
-    ..orderMileageCost = 0.0;
+    ..orderMileageCost = 0.0
+    ..foodCategory = false
+    ..clothingCategory = false
+    ..shoesCategory = false
+    ..electronicsCategory = false
+    ..jewaccessCategory = false
+    ..documentsCategory = false
+    ..otherCategory = false
+    ..orderId = ''
+    ..bookerId = ''
+    ..orderStatus = ''
+    ..reportOrderStatus = ''
+    ..reportOrderDescription = ''
+    ..reportOrderActive = false;
 
   static CollectionReference get collection =>
       FirebaseFirestore.instance.collection('bookOrder');
 
   static Stream<BookOrderRecord> getDocument(DocumentReference ref) => ref
       .snapshots()
-      .map((s) => serializers.deserializeWith(serializer, serializedData(s)));
+      .map((s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   static Future<BookOrderRecord> getDocumentOnce(DocumentReference ref) => ref
       .get()
-      .then((s) => serializers.deserializeWith(serializer, serializedData(s)));
+      .then((s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   BookOrderRecord._();
   factory BookOrderRecord([void Function(BookOrderRecordBuilder) updates]) =
@@ -112,43 +140,71 @@ abstract class BookOrderRecord
   static BookOrderRecord getDocumentFromData(
           Map<String, dynamic> data, DocumentReference reference) =>
       serializers.deserializeWith(serializer,
-          {...mapFromFirestore(data), kDocumentReferenceField: reference});
+          {...mapFromFirestore(data), kDocumentReferenceField: reference})!;
 }
 
 Map<String, dynamic> createBookOrderRecordData({
-  String senderName,
-  String senderEmail,
-  int senderPhone,
-  String senderAddress,
-  LatLng senderLocation,
-  String senderLandmark,
-  String senderDirection,
-  String receiverName,
-  String receiverEmail,
-  int receiverPhone,
-  String receiverAddress,
-  LatLng receiverLocation,
-  String receiverLandmark,
-  String receiverDirection,
-  DateTime orderTimestamp,
-  double orderMileageCost,
-}) =>
-    serializers.toFirestore(
-        BookOrderRecord.serializer,
-        BookOrderRecord((b) => b
-          ..senderName = senderName
-          ..senderEmail = senderEmail
-          ..senderPhone = senderPhone
-          ..senderAddress = senderAddress
-          ..senderLocation = senderLocation
-          ..senderLandmark = senderLandmark
-          ..senderDirection = senderDirection
-          ..receiverName = receiverName
-          ..receiverEmail = receiverEmail
-          ..receiverPhone = receiverPhone
-          ..receiverAddress = receiverAddress
-          ..receiverLocation = receiverLocation
-          ..receiverLandmark = receiverLandmark
-          ..receiverDirection = receiverDirection
-          ..orderTimestamp = orderTimestamp
-          ..orderMileageCost = orderMileageCost));
+  String? senderName,
+  String? senderEmail,
+  int? senderPhone,
+  String? senderAddress,
+  LatLng? senderLocation,
+  String? senderLandmark,
+  String? receiverName,
+  String? receiverEmail,
+  int? receiverPhone,
+  String? receiverAddress,
+  LatLng? receiverLocation,
+  String? receiverLandmark,
+  DateTime? orderTimestamp,
+  double? orderMileageCost,
+  bool? foodCategory,
+  bool? clothingCategory,
+  bool? shoesCategory,
+  bool? electronicsCategory,
+  bool? jewaccessCategory,
+  bool? documentsCategory,
+  bool? otherCategory,
+  String? orderId,
+  String? bookerId,
+  String? orderStatus,
+  String? reportOrderStatus,
+  String? reportOrderDescription,
+  bool? reportOrderActive,
+}) {
+  final firestoreData = serializers.toFirestore(
+    BookOrderRecord.serializer,
+    BookOrderRecord(
+      (b) => b
+        ..senderName = senderName
+        ..senderEmail = senderEmail
+        ..senderPhone = senderPhone
+        ..senderAddress = senderAddress
+        ..senderLocation = senderLocation
+        ..senderLandmark = senderLandmark
+        ..receiverName = receiverName
+        ..receiverEmail = receiverEmail
+        ..receiverPhone = receiverPhone
+        ..receiverAddress = receiverAddress
+        ..receiverLocation = receiverLocation
+        ..receiverLandmark = receiverLandmark
+        ..orderTimestamp = orderTimestamp
+        ..orderMileageCost = orderMileageCost
+        ..foodCategory = foodCategory
+        ..clothingCategory = clothingCategory
+        ..shoesCategory = shoesCategory
+        ..electronicsCategory = electronicsCategory
+        ..jewaccessCategory = jewaccessCategory
+        ..documentsCategory = documentsCategory
+        ..otherCategory = otherCategory
+        ..orderId = orderId
+        ..bookerId = bookerId
+        ..orderStatus = orderStatus
+        ..reportOrderStatus = reportOrderStatus
+        ..reportOrderDescription = reportOrderDescription
+        ..reportOrderActive = reportOrderActive,
+    ),
+  );
+
+  return firestoreData;
+}
