@@ -29,12 +29,13 @@ final Serializers serializers = (_$serializers.toBuilder()
       ..add(DocumentReferenceSerializer())
       ..add(DateTimeSerializer())
       ..add(LatLngSerializer())
+      ..add(FirestoreUtilDataSerializer())
       ..addPlugin(StandardJsonPlugin()))
     .build();
 
 extension SerializerExtensions on Serializers {
   Map<String, dynamic> toFirestore<T>(Serializer<T> serializer, T object) =>
-      mapToFirestore(serializeWith(serializer, object));
+      mapToFirestore(serializeWith(serializer, object) as Map<String, dynamic>);
 }
 
 class DocumentReferenceSerializer
@@ -94,24 +95,97 @@ class LatLngSerializer implements PrimitiveSerializer<LatLng> {
       serialized as LatLng;
 }
 
-Map<String, dynamic> serializedData(DocumentSnapshot doc) =>
-    {...mapFromFirestore(doc.data()), kDocumentReferenceField: doc.reference};
+class FirestoreUtilData {
+  const FirestoreUtilData({
+    this.fieldValues = const {},
+    this.clearUnsetFields = true,
+    this.create = false,
+    this.delete = false,
+  });
+  final Map<String, dynamic> fieldValues;
+  final bool clearUnsetFields;
+  final bool create;
+  final bool delete;
+  static String get name => 'firestoreUtilData';
+}
+
+class FirestoreUtilDataSerializer
+    implements PrimitiveSerializer<FirestoreUtilData> {
+  final bool structured = false;
+  @override
+  final Iterable<Type> types = new BuiltList<Type>([FirestoreUtilData]);
+  @override
+  final String wireName = 'FirestoreUtilData';
+
+  @override
+  Object serialize(Serializers serializers, FirestoreUtilData firestoreUtilData,
+      {FullType specifiedType: FullType.unspecified}) {
+    return firestoreUtilData;
+  }
+
+  @override
+  FirestoreUtilData deserialize(Serializers serializers, Object serialized,
+          {FullType specifiedType: FullType.unspecified}) =>
+      serialized as FirestoreUtilData;
+}
+
+Map<String, dynamic> serializedData(DocumentSnapshot doc) => {
+      ...mapFromFirestore(doc.data() as Map<String, dynamic>),
+      kDocumentReferenceField: doc.reference
+    };
 
 Map<String, dynamic> mapFromFirestore(Map<String, dynamic> data) =>
-    data.map((key, value) {
+    mergeNestedFields(data)
+        .where((k, _) => k != FirestoreUtilData.name)
+        .map((key, value) {
+      // Handle Timestamp
       if (value is Timestamp) {
-        value = (value as Timestamp).toDate();
+        value = value.toDate();
       }
+      // Handle list of Timestamp
+      if (value is Iterable && value.isNotEmpty && value.first is Timestamp) {
+        value = value.map((v) => (v as Timestamp).toDate()).toList();
+      }
+      // Handle GeoPoint
       if (value is GeoPoint) {
-        value = (value as GeoPoint).toLatLng();
+        value = value.toLatLng();
+      }
+      // Handle list of GeoPoint
+      if (value is Iterable && value.isNotEmpty && value.first is GeoPoint) {
+        value = value.map((v) => (v as GeoPoint).toLatLng()).toList();
+      }
+      // Handle nested data.
+      if (value is Map) {
+        value = mapFromFirestore(value as Map<String, dynamic>);
+      }
+      // Handle list of nested data.
+      if (value is Iterable && value.isNotEmpty && value.first is Map) {
+        value = value
+            .map((v) => mapFromFirestore(v as Map<String, dynamic>))
+            .toList();
       }
       return MapEntry(key, value);
     });
 
 Map<String, dynamic> mapToFirestore(Map<String, dynamic> data) =>
-    data.map((key, value) {
+    data.where((k, v) => k != FirestoreUtilData.name).map((key, value) {
+      // Handle GeoPoint
       if (value is LatLng) {
-        value = (value as LatLng).toGeoPoint();
+        value = value.toGeoPoint();
+      }
+      // Handle list of GeoPoint
+      if (value is Iterable && value.isNotEmpty && value.first is LatLng) {
+        value = value.map((v) => (v as LatLng).toGeoPoint()).toList();
+      }
+      // Handle nested data.
+      if (value is Map) {
+        value = mapFromFirestore(value as Map<String, dynamic>);
+      }
+      // Handle list of nested data.
+      if (value is Iterable && value.isNotEmpty && value.first is Map) {
+        value = value
+            .map((v) => mapFromFirestore(v as Map<String, dynamic>))
+            .toList();
       }
       return MapEntry(key, value);
     });
@@ -126,11 +200,42 @@ extension LatLngExtension on GeoPoint {
 
 DocumentReference toRef(String ref) => FirebaseFirestore.instance.doc(ref);
 
-T safeGet<T>(T Function() func, [Function(dynamic) reportError]) {
+T? safeGet<T>(T Function() func, [Function(dynamic)? reportError]) {
   try {
     return func();
   } catch (e) {
     reportError?.call(e);
   }
   return null;
+}
+
+Map<String, dynamic> mergeNestedFields(Map<String, dynamic> data) {
+  final nestedData = data.where((k, _) => k.contains('.'));
+  final fieldNames = nestedData.keys.map((k) => k.split('.').first).toSet();
+  // Remove nested values (e.g. 'foo.bar') and merge them into a map.
+  data.removeWhere((k, _) => k.contains('.'));
+  fieldNames.forEach((name) {
+    final mergedValues = mergeNestedFields(
+      nestedData
+          .where((k, _) => k.split('.').first == name)
+          .map((k, v) => MapEntry(k.split('.').skip(1).join('.'), v)),
+    );
+    final existingValue = data[name];
+    data[name] = {
+      if (existingValue != null && existingValue is Map)
+        ...existingValue as Map<String, dynamic>,
+      ...mergedValues,
+    };
+  });
+  // Merge any nested maps inside any of the fields as well.
+  data.where((_, v) => v is Map).forEach((k, v) {
+    data[k] = mergeNestedFields(v as Map<String, dynamic>);
+  });
+
+  return data;
+}
+
+extension _WhereMapExtension<K, V> on Map<K, V> {
+  Map<K, V> where(bool Function(K, V) test) =>
+      Map.fromEntries(entries.where((e) => test(e.key, e.value)));
 }
