@@ -11,33 +11,27 @@ abstract class SubscriptionRecord
   static Serializer<SubscriptionRecord> get serializer =>
       _$subscriptionRecordSerializer;
 
-  @nullable
   @BuiltValueField(wireName: 'sub_status')
-  bool get subStatus;
+  bool? get subStatus;
 
-  @nullable
   @BuiltValueField(wireName: 'sub_package')
-  DocumentReference get subPackage;
+  DocumentReference? get subPackage;
 
-  @nullable
   @BuiltValueField(wireName: 'sub_mileage_lite')
-  int get subMileageLite;
+  int? get subMileageLite;
 
-  @nullable
   @BuiltValueField(wireName: 'sub_mileage_amateur')
-  int get subMileageAmateur;
+  int? get subMileageAmateur;
 
-  @nullable
   @BuiltValueField(wireName: 'sub_mileage_professional')
-  int get subMileageProfessional;
+  int? get subMileageProfessional;
 
-  @nullable
   @BuiltValueField(wireName: 'sub_mileage_business')
-  int get subMileageBusiness;
+  int? get subMileageBusiness;
 
-  @nullable
   @BuiltValueField(wireName: kDocumentReferenceField)
-  DocumentReference get reference;
+  DocumentReference? get ffRef;
+  DocumentReference get reference => ffRef!;
 
   static void _initializeBuilder(SubscriptionRecordBuilder builder) => builder
     ..subStatus = false
@@ -51,11 +45,11 @@ abstract class SubscriptionRecord
 
   static Stream<SubscriptionRecord> getDocument(DocumentReference ref) => ref
       .snapshots()
-      .map((s) => serializers.deserializeWith(serializer, serializedData(s)));
+      .map((s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   static Future<SubscriptionRecord> getDocumentOnce(DocumentReference ref) =>
       ref.get().then(
-          (s) => serializers.deserializeWith(serializer, serializedData(s)));
+          (s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   SubscriptionRecord._();
   factory SubscriptionRecord(
@@ -65,23 +59,29 @@ abstract class SubscriptionRecord
   static SubscriptionRecord getDocumentFromData(
           Map<String, dynamic> data, DocumentReference reference) =>
       serializers.deserializeWith(serializer,
-          {...mapFromFirestore(data), kDocumentReferenceField: reference});
+          {...mapFromFirestore(data), kDocumentReferenceField: reference})!;
 }
 
 Map<String, dynamic> createSubscriptionRecordData({
-  bool subStatus,
-  DocumentReference subPackage,
-  int subMileageLite,
-  int subMileageAmateur,
-  int subMileageProfessional,
-  int subMileageBusiness,
-}) =>
-    serializers.toFirestore(
-        SubscriptionRecord.serializer,
-        SubscriptionRecord((s) => s
-          ..subStatus = subStatus
-          ..subPackage = subPackage
-          ..subMileageLite = subMileageLite
-          ..subMileageAmateur = subMileageAmateur
-          ..subMileageProfessional = subMileageProfessional
-          ..subMileageBusiness = subMileageBusiness));
+  bool? subStatus,
+  DocumentReference? subPackage,
+  int? subMileageLite,
+  int? subMileageAmateur,
+  int? subMileageProfessional,
+  int? subMileageBusiness,
+}) {
+  final firestoreData = serializers.toFirestore(
+    SubscriptionRecord.serializer,
+    SubscriptionRecord(
+      (s) => s
+        ..subStatus = subStatus
+        ..subPackage = subPackage
+        ..subMileageLite = subMileageLite
+        ..subMileageAmateur = subMileageAmateur
+        ..subMileageProfessional = subMileageProfessional
+        ..subMileageBusiness = subMileageBusiness,
+    ),
+  );
+
+  return firestoreData;
+}
