@@ -9,54 +9,64 @@ part 'users_record.g.dart';
 abstract class UsersRecord implements Built<UsersRecord, UsersRecordBuilder> {
   static Serializer<UsersRecord> get serializer => _$usersRecordSerializer;
 
-  @nullable
   @BuiltValueField(wireName: 'display_name')
-  String get displayName;
+  String? get displayName;
 
-  @nullable
-  String get email;
+  String? get email;
 
-  @nullable
-  String get password;
+  String? get password;
 
-  @nullable
-  String get uid;
+  String? get uid;
 
-  @nullable
-  int get age;
+  int? get age;
 
-  @nullable
-  LatLng get location;
+  LatLng? get location;
 
-  @nullable
   @BuiltValueField(wireName: 'phone_number')
-  String get phoneNumber;
+  String? get phoneNumber;
 
-  @nullable
   @BuiltValueField(wireName: 'photo_url')
-  String get photoUrl;
+  String? get photoUrl;
 
-  @nullable
   @BuiltValueField(wireName: 'created_time')
-  DateTime get createdTime;
+  DateTime? get createdTime;
 
-  @nullable
-  String get userTitle;
+  String? get userTitle;
 
-  @nullable
-  String get address;
+  String? get address;
 
-  @nullable
   @BuiltValueField(wireName: 'user_landmark')
-  String get userLandmark;
+  String? get userLandmark;
 
-  @nullable
   @BuiltValueField(wireName: 'bus_stop')
-  String get busStop;
+  String? get busStop;
 
-  @nullable
+  @BuiltValueField(wireName: 'mileage_balance')
+  double? get mileageBalance;
+
+  @BuiltValueField(wireName: 'mileage_cost')
+  double? get mileageCost;
+
+  @BuiltValueField(wireName: 'push_notification')
+  bool? get pushNotification;
+
+  @BuiltValueField(wireName: 'email_notification')
+  bool? get emailNotification;
+
+  @BuiltValueField(wireName: 'location_services')
+  bool? get locationServices;
+
+  @BuiltValueField(wireName: 'first_name')
+  String? get firstName;
+
+  @BuiltValueField(wireName: 'last_name')
+  String? get lastName;
+
+  String? get gender;
+
   @BuiltValueField(wireName: kDocumentReferenceField)
-  DocumentReference get reference;
+  DocumentReference? get ffRef;
+  DocumentReference get reference => ffRef!;
 
   static void _initializeBuilder(UsersRecordBuilder builder) => builder
     ..displayName = ''
@@ -69,18 +79,26 @@ abstract class UsersRecord implements Built<UsersRecord, UsersRecordBuilder> {
     ..userTitle = ''
     ..address = ''
     ..userLandmark = ''
-    ..busStop = '';
+    ..busStop = ''
+    ..mileageBalance = 0.0
+    ..mileageCost = 0.0
+    ..pushNotification = false
+    ..emailNotification = false
+    ..locationServices = false
+    ..firstName = ''
+    ..lastName = ''
+    ..gender = '';
 
   static CollectionReference get collection =>
       FirebaseFirestore.instance.collection('users');
 
   static Stream<UsersRecord> getDocument(DocumentReference ref) => ref
       .snapshots()
-      .map((s) => serializers.deserializeWith(serializer, serializedData(s)));
+      .map((s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   static Future<UsersRecord> getDocumentOnce(DocumentReference ref) => ref
       .get()
-      .then((s) => serializers.deserializeWith(serializer, serializedData(s)));
+      .then((s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   UsersRecord._();
   factory UsersRecord([void Function(UsersRecordBuilder) updates]) =
@@ -89,37 +107,59 @@ abstract class UsersRecord implements Built<UsersRecord, UsersRecordBuilder> {
   static UsersRecord getDocumentFromData(
           Map<String, dynamic> data, DocumentReference reference) =>
       serializers.deserializeWith(serializer,
-          {...mapFromFirestore(data), kDocumentReferenceField: reference});
+          {...mapFromFirestore(data), kDocumentReferenceField: reference})!;
 }
 
 Map<String, dynamic> createUsersRecordData({
-  String displayName,
-  String email,
-  String password,
-  String uid,
-  int age,
-  LatLng location,
-  String phoneNumber,
-  String photoUrl,
-  DateTime createdTime,
-  String userTitle,
-  String address,
-  String userLandmark,
-  String busStop,
-}) =>
-    serializers.toFirestore(
-        UsersRecord.serializer,
-        UsersRecord((u) => u
-          ..displayName = displayName
-          ..email = email
-          ..password = password
-          ..uid = uid
-          ..age = age
-          ..location = location
-          ..phoneNumber = phoneNumber
-          ..photoUrl = photoUrl
-          ..createdTime = createdTime
-          ..userTitle = userTitle
-          ..address = address
-          ..userLandmark = userLandmark
-          ..busStop = busStop));
+  String? displayName,
+  String? email,
+  String? password,
+  String? uid,
+  int? age,
+  LatLng? location,
+  String? phoneNumber,
+  String? photoUrl,
+  DateTime? createdTime,
+  String? userTitle,
+  String? address,
+  String? userLandmark,
+  String? busStop,
+  double? mileageBalance,
+  double? mileageCost,
+  bool? pushNotification,
+  bool? emailNotification,
+  bool? locationServices,
+  String? firstName,
+  String? lastName,
+  String? gender,
+}) {
+  final firestoreData = serializers.toFirestore(
+    UsersRecord.serializer,
+    UsersRecord(
+      (u) => u
+        ..displayName = displayName
+        ..email = email
+        ..password = password
+        ..uid = uid
+        ..age = age
+        ..location = location
+        ..phoneNumber = phoneNumber
+        ..photoUrl = photoUrl
+        ..createdTime = createdTime
+        ..userTitle = userTitle
+        ..address = address
+        ..userLandmark = userLandmark
+        ..busStop = busStop
+        ..mileageBalance = mileageBalance
+        ..mileageCost = mileageCost
+        ..pushNotification = pushNotification
+        ..emailNotification = emailNotification
+        ..locationServices = locationServices
+        ..firstName = firstName
+        ..lastName = lastName
+        ..gender = gender,
+    ),
+  );
+
+  return firestoreData;
+}
