@@ -1,60 +1,133 @@
+// ignore_for_file: overridden_fields, annotate_overrides
+
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
-// ignore: avoid_classes_with_only_static_members
-class FlutterFlowTheme {
-  static const Color primaryColor = Color(0xFFFFCD3C);
-  static const Color secondaryColor = Color(0xFFFFDA6F);
-  static const Color tertiaryColor = Color(0xFFFFCD3C);
+abstract class FlutterFlowTheme {
+  static FlutterFlowTheme of(BuildContext context) {
+    return LightModeTheme();
+  }
 
-  static const Color background = Color(0xFF1A1F24);
-  static const Color darkBackground = Color(0xFF111417);
-  static const Color textColor = Color(0xFFFFFFFF);
-  static const Color grayDark = Color(0xFF57636C);
-  static const Color grayLight = Color(0xFF8B97A2);
-  static const Color errorRed = Color(0xFFF06A6A);
+  late Color primaryColor;
+  late Color secondaryColor;
+  late Color tertiaryColor;
+  late Color alternate;
+  late Color primaryBackground;
+  late Color secondaryBackground;
+  late Color primaryText;
+  late Color secondaryText;
 
-  String primaryFontFamily = 'Poppins';
-  String secondaryFontFamily = 'Roboto';
-  static TextStyle get title1 => GoogleFonts.getFont(
+  late Color background;
+  late Color darkBackground;
+  late Color textColor;
+  late Color grayDark;
+  late Color grayLight;
+  late Color errorRed;
+
+  String get title1Family => typography.title1Family;
+  TextStyle get title1 => typography.title1;
+  String get title2Family => typography.title2Family;
+  TextStyle get title2 => typography.title2;
+  String get title3Family => typography.title3Family;
+  TextStyle get title3 => typography.title3;
+  String get subtitle1Family => typography.subtitle1Family;
+  TextStyle get subtitle1 => typography.subtitle1;
+  String get subtitle2Family => typography.subtitle2Family;
+  TextStyle get subtitle2 => typography.subtitle2;
+  String get bodyText1Family => typography.bodyText1Family;
+  TextStyle get bodyText1 => typography.bodyText1;
+  String get bodyText2Family => typography.bodyText2Family;
+  TextStyle get bodyText2 => typography.bodyText2;
+
+  Typography get typography => ThemeTypography(this);
+}
+
+class LightModeTheme extends FlutterFlowTheme {
+  late Color primaryColor = const Color(0xFFFFCD3C);
+  late Color secondaryColor = const Color(0xFFFFDA6F);
+  late Color tertiaryColor = const Color(0xFFFFCD3C);
+  late Color alternate = const Color(0x00000000);
+  late Color primaryBackground = const Color(0x00000000);
+  late Color secondaryBackground = const Color(0x00000000);
+  late Color primaryText = const Color(0xFFFFFFFF);
+  late Color secondaryText = const Color(0xFFFFFFFF);
+
+  late Color background = Color(0xFF1A1F24);
+  late Color darkBackground = Color(0xFF111417);
+  late Color textColor = Color(0xFFFFFFFF);
+  late Color grayDark = Color(0xFF57636C);
+  late Color grayLight = Color(0xFF8B97A2);
+  late Color errorRed = Color(0xFFF06A6A);
+}
+
+abstract class Typography {
+  String get title1Family;
+  TextStyle get title1;
+  String get title2Family;
+  TextStyle get title2;
+  String get title3Family;
+  TextStyle get title3;
+  String get subtitle1Family;
+  TextStyle get subtitle1;
+  String get subtitle2Family;
+  TextStyle get subtitle2;
+  String get bodyText1Family;
+  TextStyle get bodyText1;
+  String get bodyText2Family;
+  TextStyle get bodyText2;
+}
+
+class ThemeTypography extends Typography {
+  ThemeTypography(this.theme);
+
+  final FlutterFlowTheme theme;
+
+  String get title1Family => 'Lexend Deca';
+  TextStyle get title1 => GoogleFonts.getFont(
         'Lexend Deca',
-        color: textColor,
+        color: theme.textColor,
         fontWeight: FontWeight.bold,
         fontSize: 24,
       );
-  static TextStyle get title2 => GoogleFonts.getFont(
+  String get title2Family => 'Lexend Deca';
+  TextStyle get title2 => GoogleFonts.getFont(
         'Lexend Deca',
-        color: primaryColor,
+        color: theme.primaryColor,
         fontWeight: FontWeight.w500,
         fontSize: 28,
       );
-  static TextStyle get title3 => GoogleFonts.getFont(
+  String get title3Family => 'Lexend Deca';
+  TextStyle get title3 => GoogleFonts.getFont(
         'Lexend Deca',
-        color: textColor,
+        color: theme.textColor,
         fontWeight: FontWeight.w500,
         fontSize: 20,
       );
-  static TextStyle get subtitle1 => GoogleFonts.getFont(
+  String get subtitle1Family => 'Lexend Deca';
+  TextStyle get subtitle1 => GoogleFonts.getFont(
         'Lexend Deca',
-        color: grayLight,
+        color: theme.grayLight,
         fontWeight: FontWeight.w500,
         fontSize: 18,
       );
-  static TextStyle get subtitle2 => GoogleFonts.getFont(
+  String get subtitle2Family => 'Lexend Deca';
+  TextStyle get subtitle2 => GoogleFonts.getFont(
         'Lexend Deca',
-        color: grayLight,
+        color: theme.grayLight,
         fontWeight: FontWeight.normal,
         fontSize: 16,
       );
-  static TextStyle get bodyText1 => GoogleFonts.getFont(
+  String get bodyText1Family => 'Lexend Deca';
+  TextStyle get bodyText1 => GoogleFonts.getFont(
         'Lexend Deca',
-        color: grayLight,
+        color: theme.grayLight,
         fontWeight: FontWeight.normal,
         fontSize: 14,
       );
-  static TextStyle get bodyText2 => GoogleFonts.getFont(
+  String get bodyText2Family => 'Lexend Deca';
+  TextStyle get bodyText2 => GoogleFonts.getFont(
         'Lexend Deca',
-        color: textColor,
+        color: theme.textColor,
         fontWeight: FontWeight.normal,
         fontSize: 14,
       );
@@ -62,29 +135,35 @@ class FlutterFlowTheme {
 
 extension TextStyleHelper on TextStyle {
   TextStyle override({
-    String fontFamily,
-    Color color,
-    double fontSize,
-    FontWeight fontWeight,
-    FontStyle fontStyle,
+    String? fontFamily,
+    Color? color,
+    double? fontSize,
+    FontWeight? fontWeight,
+    double? letterSpacing,
+    FontStyle? fontStyle,
     bool useGoogleFonts = true,
-    double lineHeight,
+    TextDecoration? decoration,
+    double? lineHeight,
   }) =>
       useGoogleFonts
           ? GoogleFonts.getFont(
-              fontFamily,
+              fontFamily!,
               color: color ?? this.color,
               fontSize: fontSize ?? this.fontSize,
+              letterSpacing: letterSpacing ?? this.letterSpacing,
               fontWeight: fontWeight ?? this.fontWeight,
               fontStyle: fontStyle ?? this.fontStyle,
+              decoration: decoration,
               height: lineHeight,
             )
           : copyWith(
               fontFamily: fontFamily,
               color: color,
               fontSize: fontSize,
+              letterSpacing: letterSpacing,
               fontWeight: fontWeight,
               fontStyle: fontStyle,
+              decoration: decoration,
               height: lineHeight,
             );
 }
